refactor(htmx): extract deal/sourcing cluster (buy_plans + offers + sourcing) from htmx_views
Continues the htmx_views.py domain split. Moves the deal/sourcing cluster
verbatim into three new sub-routers under app/routers/htmx/, each with its
own APIRouter(tags=["htmx-views"]) and unchanged /v2 paths, mounted in main.py:

- buy_plans.py — Approvals (Buy Plans) hub: /v2/partials/approvals*,
  sales-orders new/create, the buy-plans resource/orders/board/archive/
  supervise boards, per-plan lifecycle actions (submit/approve/verify-so/
  confirm-po/resource/claim/verify-po/issue/cancel/reset), and the legacy
  /v2/buy-plans 302 redirect. Carries its cluster-private helpers
  (_default_lens, _can_supervise/_can_resource/_can_see_all_deals/
  _resolve_deal_scope/_require_po_cutter, _render_supervise_body,
  _APPROVALS_TABS/_TAB_APPROVE_ATTR/_PO_CUTTER_ROLES). The interleaved
  settings/ops-group|users|scorecard routes stay in htmx_views (settings domain).
- offers.py — AI offer parsing, offer CRUD + review/promote/reject/changelog,
  quote-from-offers, activity logging, RFQ compose/cleanup/rephrase/send,
  follow-ups, and vendor-response review/reply. The interleaved requisition-
  management routes (bulk/inline-edit/row-action/delete/update-requirement/
  poll-inbox) stay in htmx_views.
- sourcing.py — self-contained /v2/sourcing/* + /v2/partials/sourcing/*.

Shared helpers _safe_int/_safe_float/_is_ops_member moved to _shared.py
(each shared between a moved cluster and a staying route); htmx_views
re-imports _safe_int/_is_ops_member for its remaining routes.

Route parity IDENTICAL (sorted method+path dump of app.routes — empty diff,
770 routes). Full suite green (21101 passed). pre-commit --all-files clean.
Retargeted tests that imported/patched moved symbols to the new modules.
APP_MAP_ARCHITECTURE updated.
