+++ app/main.py
@@ -688,8 +688,11 @@ from .routers.crm import router as crm_router
 from .routers.documents import router as documents_router
 from .routers.error_reports import router as error_reports_router
 from .routers.events import router as events_router
+from .routers.htmx.buy_plans import router as htmx_buy_plans_router
 from .routers.htmx.companies import router as htmx_companies_router
+from .routers.htmx.offers import router as htmx_offers_router
 from .routers.htmx.requisitions import router as htmx_requisitions_router
+from .routers.htmx.sourcing import router as htmx_sourcing_router
 from .routers.htmx.vendors import router as htmx_vendors_router
 from .routers.htmx_views import router as htmx_views_router
 from .routers.materials import router as materials_router
@@ -741,3 +744,6 @@ app.include_router(htmx_views_router)
 app.include_router(htmx_requisitions_router)
 app.include_router(htmx_vendors_router)
 app.include_router(htmx_companies_router)
+app.include_router(htmx_buy_plans_router)
+app.include_router(htmx_offers_router)
+app.include_router(htmx_sourcing_router)
