+++ app/routers/htmx/sourcing.py
@@ -1,0 +1,794 @@
+"""routers/htmx/sourcing.py — Sourcing engine partial views (HTMX + Alpine).
+
+Server-rendered HTML partials + SSE streams for the sourcing surface: the results
+page/stream, manual search trigger, lead detail/status/feedback, and the split-panel
+workspace (page, list, lead panel). Self-contained slice extracted verbatim from
+htmx_views.py (same `/v2/sourcing` + `/v2/partials/sourcing` paths, same `htmx-views`
+tag).
+
+Called by: app/main.py (router mount).
+Depends on: app.models, app.dependencies, app.database, app.scoring,
+    app.search_service, ._shared.
+"""
+
+import asyncio
+import json
+import time
+from datetime import datetime, timedelta, timezone
+
+from fastapi import APIRouter, Depends, HTTPException, Query, Request
+from fastapi.responses import HTMLResponse
+from loguru import logger
+from sqlalchemy.orm import Session
+
+from ...database import get_db
+from ...dependencies import (
+    get_user,
+    require_requisition_access,
+    require_user,
+)
+from ...models import (
+    Requirement,
+    Sighting,
+    User,
+)
+from ...template_env import page_response, template_response, templates
+from ..auth import _password_login_enabled
+from ._shared import _base_ctx, _vite_assets
+
+router = APIRouter(tags=["htmx-views"])
+
+
+@router.get("/v2/sourcing/{requirement_id}", response_class=HTMLResponse)
+async def v2_sourcing_page(request: Request, requirement_id: int, db: Session = Depends(get_db)):
+    """Full page load for sourcing results."""
+    user = get_user(request, db)
+    if not user:
+        return template_response(
+            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
+        )
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx["partial_url"] = f"/v2/partials/sourcing/{requirement_id}"
+    return page_response(ctx)
+
+
+@router.get("/v2/sourcing/leads/{lead_id}", response_class=HTMLResponse)
+async def v2_lead_detail_page(request: Request, lead_id: int, db: Session = Depends(get_db)):
+    """Full page load for lead detail."""
+    user = get_user(request, db)
+    if not user:
+        return template_response(
+            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
+        )
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx["partial_url"] = f"/v2/partials/sourcing/leads/{lead_id}"
+    return page_response(ctx)
+
+
+@router.get("/v2/partials/sourcing/{requirement_id}/stream")
+async def sourcing_stream(
+    request: Request,
+    requirement_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """SSE endpoint for sourcing search progress.
+
+    Streams per-source completion events as connectors finish searching.
+    Client connects via hx-ext="sse" sse-connect attribute.
+    Channel: sourcing:{requirement_id}
+    """
+    from sse_starlette.sse import EventSourceResponse
+
+    from ...services.sse_broker import broker
+
+    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
+    if not req:
+        raise HTTPException(404, "Requirement not found")
+
+    async def event_generator():
+        async for msg in broker.listen(f"sourcing:{requirement_id}"):
+            if await request.is_disconnected():
+                break
+            yield {
+                "event": msg["event"],
+                "data": msg["data"],
+            }
+
+    return EventSourceResponse(event_generator())
+
+
+@router.post("/v2/partials/sourcing/{requirement_id}/search", response_class=HTMLResponse)
+async def sourcing_search_trigger(
+    request: Request,
+    requirement_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Trigger multi-source search for a requirement.
+
+    Runs connectors in parallel, publishes SSE events per source completion, syncs leads
+    on completion, returns redirect to sourcing results.
+    """
+
+    from ...services.sse_broker import broker
+
+    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
+    if not req:
+        raise HTTPException(404, "Requirement not found")
+    # Search triggers connector SPEND + cross-owner disclosure — scope to the owner.
+    require_requisition_access(db, req.requisition_id, user, label="Requirement")
+
+    mpn = req.primary_mpn or ""
+    sources = ["brokerbin", "nexar", "digikey", "mouser", "oemsecrets", "element14"]
+    channel = f"sourcing:{requirement_id}"
+    all_sightings = []
+
+    async def search_source(source_name):
+        start_t = time.time()
+        try:
+            from ...search_service import quick_search_mpn
+
+            raw = await quick_search_mpn(mpn, db)
+            results = raw if isinstance(raw, list) else raw.get("sightings", [])
+            elapsed = int((time.time() - start_t) * 1000)
+            count = len(results) if results else 0
+            await broker.publish(
+                channel,
+                "source-complete",
+                json.dumps({"source": source_name, "count": count, "elapsed_ms": elapsed, "status": "done"}),
+            )
+            return results or []
+        except Exception as exc:
+            elapsed = int((time.time() - start_t) * 1000)
+            logger.error("Sourcing search failed for {} on {}: {}", mpn, source_name, exc)
+            await broker.publish(
+                channel,
+                "source-complete",
+                json.dumps(
+                    {"source": source_name, "count": 0, "elapsed_ms": elapsed, "status": "failed", "error": str(exc)}
+                ),
+            )
+            return []
+
+    results_by_source = await asyncio.gather(*[search_source(s) for s in sources], return_exceptions=True)
+
+    for source_results in results_by_source:
+        if isinstance(source_results, list):
+            all_sightings.extend(source_results)
+
+    await broker.publish(
+        channel, "search-complete", json.dumps({"total": len(all_sightings), "requirement_id": requirement_id})
+    )
+
+    return HTMLResponse(status_code=200, headers={"HX-Redirect": f"/v2/sourcing/{requirement_id}"})
+
+
+@router.get("/v2/partials/sourcing/{requirement_id}", response_class=HTMLResponse)
+async def sourcing_results_partial(
+    request: Request,
+    requirement_id: int,
+    confidence: str = "",
+    safety: str = "",
+    freshness: str = "",
+    source: str = "",
+    status: str = "",
+    contactability: str = "",
+    corroborated: str = "",
+    sort: str = "best",
+    page: int = Query(1, ge=1),
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Return sourcing results with lead cards for a requirement.
+
+    Supports filtering by confidence band, safety band, freshness window, source type,
+    buyer status, contactability, and corroboration. Sorts by best overall (default),
+    freshest, safest, easiest to contact, or most proven.
+    """
+    from ...models.sourcing_lead import SourcingLead
+
+    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
+    if not req:
+        raise HTTPException(404, "Requirement not found")
+
+    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
+
+    if confidence:
+        bands = [b.strip() for b in confidence.split(",")]
+        query = query.filter(SourcingLead.confidence_band.in_(bands))
+    if safety:
+        bands = [b.strip() for b in safety.split(",")]
+        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
+    if freshness and freshness != "all":
+        now = datetime.now(timezone.utc)
+        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
+        if freshness in cutoffs:
+            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
+    if source:
+        sources_list = [s.strip() for s in source.split(",")]
+        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
+    if status and status != "all":
+        statuses = [s.strip() for s in status.split(",")]
+        query = query.filter(SourcingLead.buyer_status.in_(statuses))
+    if contactability == "has_email":
+        query = query.filter(SourcingLead.contact_email.isnot(None))
+    elif contactability == "has_phone":
+        query = query.filter(SourcingLead.contact_phone.isnot(None))
+    if corroborated == "yes":
+        query = query.filter(SourcingLead.corroborated.is_(True))
+    elif corroborated == "no":
+        query = query.filter(SourcingLead.corroborated.is_(False))
+
+    sort_map = {
+        "best": [SourcingLead.confidence_score.desc()],
+        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
+        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
+        "contact": [SourcingLead.contactability_score.desc().nullslast()],
+        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
+    }
+    for col in sort_map.get(sort, sort_map["best"]):
+        query = query.order_by(col)
+
+    total = query.count()
+    per_page = 24
+    leads = query.offset((page - 1) * per_page).limit(per_page).all()
+
+    lead_sighting_data = {}
+    if leads:
+        for lead in leads:
+            best_sighting = (
+                db.query(Sighting)
+                .filter(
+                    Sighting.requirement_id == requirement_id,
+                    Sighting.vendor_name_normalized == lead.vendor_name_normalized,
+                )
+                .order_by(Sighting.created_at.desc().nullslast())
+                .first()
+            )
+            if best_sighting:
+                lead_sighting_data[lead.id] = {
+                    "qty_available": best_sighting.qty_available,
+                    "unit_price": best_sighting.unit_price,
+                }
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update(
+        {
+            "requirement": req,
+            "leads": leads,
+            "lead_sighting_data": lead_sighting_data,
+            "total": total,
+            "page": page,
+            "total_pages": max(1, (total + per_page - 1) // per_page),
+            "per_page": per_page,
+            "f_confidence": confidence,
+            "f_safety": safety,
+            "f_freshness": freshness,
+            "f_source": source,
+            "f_status": status,
+            "f_contactability": contactability,
+            "f_corroborated": corroborated,
+            "f_sort": sort,
+        }
+    )
+    return template_response("htmx/partials/sourcing/results.html", ctx)
+
+
+@router.get("/v2/partials/sourcing/leads/{lead_id}", response_class=HTMLResponse)
+async def lead_detail_partial(
+    request: Request,
+    lead_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Return lead detail as HTML partial.
+
+    Loads the SourcingLead, its evidence (sorted by confidence_impact desc), groups
+    evidence by source category, fetches vendor card and best sighting.
+    """
+    from ...models.sourcing_lead import LeadEvidence, SourcingLead
+    from ...services.sourcing_leads import _source_category
+
+    lead = db.query(SourcingLead).filter(SourcingLead.id == lead_id).first()
+    if not lead:
+        raise HTTPException(404, "Lead not found")
+
+    evidence = (
+        db.query(LeadEvidence)
+        .filter(LeadEvidence.lead_id == lead.id)
+        .order_by(LeadEvidence.confidence_impact.desc().nullslast())
+        .all()
+    )
+
+    evidence_by_category = {}
+    for ev in evidence:
+        cat = _source_category(ev.source_type)
+        evidence_by_category.setdefault(cat, []).append(ev)
+
+    category_labels = {
+        "api": "API",
+        "marketplace": "Marketplace",
+        "salesforce_history": "Salesforce History",
+        "avail_history": "Avail History",
+        "web_ai": "Web / AI",
+        "safety_review": "Safety Review",
+        "buyer_feedback": "Buyer Feedback",
+    }
+
+    requirement = db.query(Requirement).filter(Requirement.id == lead.requirement_id).first()
+
+    vendor_card = lead.vendor_card
+
+    best_sighting = (
+        db.query(Sighting)
+        .filter(
+            Sighting.requirement_id == lead.requirement_id,
+            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
+        )
+        .order_by(Sighting.created_at.desc().nullslast())
+        .first()
+    )
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update(
+        {
+            "lead": lead,
+            "evidence": evidence,
+            "evidence_by_category": evidence_by_category,
+            "category_labels": category_labels,
+            "requirement": requirement,
+            "vendor_card": vendor_card,
+            "best_sighting": best_sighting,
+        }
+    )
+    return template_response("htmx/partials/sourcing/lead_detail.html", ctx)
+
+
+@router.post("/v2/partials/sourcing/leads/{lead_id}/status", response_class=HTMLResponse)
+async def lead_status_update(
+    request: Request,
+    lead_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Update lead buyer status.
+
+    Returns updated lead card when called from results view (for OOB swap), or updated
+    lead detail when called from lead detail page.
+    """
+    from ...models.sourcing_lead import SourcingLead
+    from ...services.sourcing_leads import update_lead_status
+
+    _lead = db.get(SourcingLead, lead_id)
+    if not _lead:
+        raise HTTPException(404, "Lead not found")
+    require_requisition_access(db, _lead.requisition_id, user, label="Lead")
+
+    form = await request.form()
+    status_val = form.get("status", "").strip()
+    note = form.get("note", "").strip() or None
+
+    try:
+        lead = update_lead_status(
+            db,
+            lead_id,
+            status_val,
+            note=note,
+            actor_user_id=user.id,
+        )
+    except ValueError as e:
+        raise HTTPException(400, str(e))
+
+    if not lead:
+        raise HTTPException(404, "Lead not found")
+
+    hx_target = request.headers.get("HX-Target", "")
+    referer = request.headers.get("HX-Current-URL", "")
+
+    # Workspace panel context: return updated panel detail
+    if hx_target == "split-right-sourcing":
+        return await lead_panel_partial(request, lead_id, user, db)
+
+    # Full-page lead detail context
+    if "/leads/" in referer:
+        return await lead_detail_partial(request, lead_id, user, db)
+
+    # Workspace lead row context: return updated lead row
+    if hx_target.startswith("lead-row-"):
+        best_sighting = (
+            db.query(Sighting)
+            .filter(
+                Sighting.requirement_id == lead.requirement_id,
+                Sighting.vendor_name_normalized == lead.vendor_name_normalized,
+            )
+            .order_by(Sighting.created_at.desc().nullslast())
+            .first()
+        )
+        lead_sighting_data = {}
+        if best_sighting:
+            lead_sighting_data[lead.id] = {
+                "qty_available": best_sighting.qty_available,
+                "unit_price": best_sighting.unit_price,
+            }
+        ctx = _base_ctx(request, user, "requisitions")
+        ctx.update({"lead": lead, "lead_sighting_data": lead_sighting_data, "selected_lead_id": 0})
+        return template_response("htmx/partials/sourcing/lead_row.html", ctx)
+
+    # Default: card view (results grid)
+    best_sighting = (
+        db.query(Sighting)
+        .filter(
+            Sighting.requirement_id == lead.requirement_id,
+            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
+        )
+        .order_by(Sighting.created_at.desc().nullslast())
+        .first()
+    )
+    lead_sighting_data = {}
+    if best_sighting:
+        lead_sighting_data[lead.id] = {
+            "qty_available": best_sighting.qty_available,
+            "unit_price": best_sighting.unit_price,
+        }
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update({"lead": lead, "lead_sighting_data": lead_sighting_data})
+    return template_response("htmx/partials/sourcing/lead_card.html", ctx)
+
+
+@router.post("/v2/partials/sourcing/leads/{lead_id}/feedback", response_class=HTMLResponse)
+async def lead_feedback(
+    request: Request,
+    lead_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Add feedback event to a lead without changing status.
+
+    Returns updated lead detail.
+    """
+    from ...models.sourcing_lead import SourcingLead
+    from ...services.sourcing_leads import append_lead_feedback
+
+    _lead = db.get(SourcingLead, lead_id)
+    if not _lead:
+        raise HTTPException(404, "Lead not found")
+    require_requisition_access(db, _lead.requisition_id, user, label="Lead")
+
+    form = await request.form()
+    note = form.get("note", "").strip() or None
+    reason_code = form.get("reason_code", "").strip() or None
+    contact_method = form.get("contact_method", "").strip() or None
+
+    lead = append_lead_feedback(
+        db,
+        lead_id,
+        note=note,
+        reason_code=reason_code,
+        contact_method=contact_method,
+        actor_user_id=user.id,
+    )
+    if not lead:
+        raise HTTPException(404, "Lead not found")
+
+    return await lead_detail_partial(request, lead_id, user, db)
+
+
+@router.get("/v2/sourcing/{requirement_id}/workspace", response_class=HTMLResponse)
+async def v2_sourcing_workspace_page(request: Request, requirement_id: int, db: Session = Depends(get_db)):
+    """Full page load for sourcing workspace (split-panel view)."""
+    user = get_user(request, db)
+    if not user:
+        return template_response(
+            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
+        )
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx["partial_url"] = f"/v2/partials/sourcing/{requirement_id}/workspace"
+    return page_response(ctx)
+
+
+@router.get("/v2/partials/sourcing/{requirement_id}/workspace", response_class=HTMLResponse)
+async def sourcing_workspace_partial(
+    request: Request,
+    requirement_id: int,
+    confidence: str = "",
+    safety: str = "",
+    freshness: str = "",
+    source: str = "",
+    status: str = "",
+    contactability: str = "",
+    corroborated: str = "",
+    sort: str = "best",
+    page: int = Query(1, ge=1),
+    lead: int = Query(0, ge=0),
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Return sourcing workspace split-panel layout.
+
+    Reuses the same filtering/sorting logic as sourcing_results_partial but renders lead
+    rows in a split-panel instead of card grid. Optional lead=ID param pre-selects a
+    lead in the right panel.
+    """
+    from ...models.sourcing_lead import SourcingLead
+
+    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
+    if not req:
+        raise HTTPException(404, "Requirement not found")
+
+    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
+
+    if confidence:
+        bands = [b.strip() for b in confidence.split(",")]
+        query = query.filter(SourcingLead.confidence_band.in_(bands))
+    if safety:
+        bands = [b.strip() for b in safety.split(",")]
+        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
+    if freshness and freshness != "all":
+        now = datetime.now(timezone.utc)
+        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
+        if freshness in cutoffs:
+            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
+    if source:
+        sources_list = [s.strip() for s in source.split(",")]
+        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
+    if status and status != "all":
+        statuses = [s.strip() for s in status.split(",")]
+        query = query.filter(SourcingLead.buyer_status.in_(statuses))
+    if contactability == "has_email":
+        query = query.filter(SourcingLead.contact_email.isnot(None))
+    elif contactability == "has_phone":
+        query = query.filter(SourcingLead.contact_phone.isnot(None))
+    if corroborated == "yes":
+        query = query.filter(SourcingLead.corroborated.is_(True))
+    elif corroborated == "no":
+        query = query.filter(SourcingLead.corroborated.is_(False))
+
+    sort_map = {
+        "best": [SourcingLead.confidence_score.desc()],
+        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
+        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
+        "contact": [SourcingLead.contactability_score.desc().nullslast()],
+        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
+    }
+    for col in sort_map.get(sort, sort_map["best"]):
+        query = query.order_by(col)
+
+    total = query.count()
+    per_page = 24
+    leads = query.offset((page - 1) * per_page).limit(per_page).all()
+
+    lead_sighting_data = {}
+    if leads:
+        for ld in leads:
+            best_sighting = (
+                db.query(Sighting)
+                .filter(
+                    Sighting.requirement_id == requirement_id,
+                    Sighting.vendor_name_normalized == ld.vendor_name_normalized,
+                )
+                .order_by(Sighting.created_at.desc().nullslast())
+                .first()
+            )
+            if best_sighting:
+                lead_sighting_data[ld.id] = {
+                    "qty_available": best_sighting.qty_available,
+                    "unit_price": best_sighting.unit_price,
+                }
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update(
+        {
+            "requirement": req,
+            "leads": leads,
+            "lead_sighting_data": lead_sighting_data,
+            "total": total,
+            "page": page,
+            "total_pages": max(1, (total + per_page - 1) // per_page),
+            "per_page": per_page,
+            "f_confidence": confidence,
+            "f_safety": safety,
+            "f_freshness": freshness,
+            "f_source": source,
+            "f_status": status,
+            "f_contactability": contactability,
+            "f_corroborated": corroborated,
+            "f_sort": sort,
+            "selected_lead_id": lead if lead else 0,
+        }
+    )
+    return template_response("htmx/partials/sourcing/workspace.html", ctx)
+
+
+@router.get("/v2/partials/sourcing/{requirement_id}/workspace-list", response_class=HTMLResponse)
+async def sourcing_workspace_list_partial(
+    request: Request,
+    requirement_id: int,
+    confidence: str = "",
+    safety: str = "",
+    freshness: str = "",
+    source: str = "",
+    status: str = "",
+    contactability: str = "",
+    corroborated: str = "",
+    sort: str = "best",
+    page: int = Query(1, ge=1),
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Return just the lead list rows for the workspace left panel.
+
+    Used when filters change — swaps only #lead-list-content without touching the right
+    panel or overall layout.
+    """
+    from ...models.sourcing_lead import SourcingLead
+
+    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
+    if not req:
+        raise HTTPException(404, "Requirement not found")
+
+    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
+
+    if confidence:
+        bands = [b.strip() for b in confidence.split(",")]
+        query = query.filter(SourcingLead.confidence_band.in_(bands))
+    if safety:
+        bands = [b.strip() for b in safety.split(",")]
+        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
+    if freshness and freshness != "all":
+        now = datetime.now(timezone.utc)
+        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
+        if freshness in cutoffs:
+            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
+    if source:
+        sources_list = [s.strip() for s in source.split(",")]
+        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
+    if status and status != "all":
+        statuses = [s.strip() for s in status.split(",")]
+        query = query.filter(SourcingLead.buyer_status.in_(statuses))
+    if contactability == "has_email":
+        query = query.filter(SourcingLead.contact_email.isnot(None))
+    elif contactability == "has_phone":
+        query = query.filter(SourcingLead.contact_phone.isnot(None))
+    if corroborated == "yes":
+        query = query.filter(SourcingLead.corroborated.is_(True))
+    elif corroborated == "no":
+        query = query.filter(SourcingLead.corroborated.is_(False))
+
+    sort_map = {
+        "best": [SourcingLead.confidence_score.desc()],
+        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
+        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
+        "contact": [SourcingLead.contactability_score.desc().nullslast()],
+        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
+    }
+    for col in sort_map.get(sort, sort_map["best"]):
+        query = query.order_by(col)
+
+    total = query.count()
+    per_page = 24
+    leads = query.offset((page - 1) * per_page).limit(per_page).all()
+
+    lead_sighting_data = {}
+    if leads:
+        for ld in leads:
+            best_sighting = (
+                db.query(Sighting)
+                .filter(
+                    Sighting.requirement_id == requirement_id,
+                    Sighting.vendor_name_normalized == ld.vendor_name_normalized,
+                )
+                .order_by(Sighting.created_at.desc().nullslast())
+                .first()
+            )
+            if best_sighting:
+                lead_sighting_data[ld.id] = {
+                    "qty_available": best_sighting.qty_available,
+                    "unit_price": best_sighting.unit_price,
+                }
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update(
+        {
+            "requirement": req,
+            "leads": leads,
+            "lead_sighting_data": lead_sighting_data,
+            "total": total,
+            "page": page,
+            "total_pages": max(1, (total + per_page - 1) // per_page),
+            "per_page": per_page,
+            "f_confidence": confidence,
+            "f_safety": safety,
+            "f_freshness": freshness,
+            "f_source": source,
+            "f_status": status,
+            "f_contactability": contactability,
+            "f_corroborated": corroborated,
+            "f_sort": sort,
+            "selected_lead_id": 0,
+        }
+    )
+
+    html_parts = []
+    if leads:
+        for ld in leads:
+            rendered = templates.get_template("htmx/partials/sourcing/lead_row.html").render({**ctx, "lead": ld})
+            html_parts.append(rendered)
+    else:
+        html_parts.append(
+            '<div class="flex flex-col items-center justify-center py-12 text-gray-400">'
+            '<p class="text-sm">No leads found</p></div>'
+        )
+
+    return HTMLResponse("".join(html_parts))
+
+
+@router.get("/v2/partials/sourcing/leads/{lead_id}/panel", response_class=HTMLResponse)
+async def lead_panel_partial(
+    request: Request,
+    lead_id: int,
+    user: User = Depends(require_user),
+    db: Session = Depends(get_db),
+):
+    """Return condensed lead detail for workspace right panel.
+
+    Same data as lead_detail_partial but uses lead_panel.html template (no breadcrumb,
+    collapsible sections, denser layout). Includes OOB swap of the lead row in the left
+    panel for highlight update.
+    """
+    from ...models.sourcing_lead import LeadEvidence, SourcingLead
+    from ...services.sourcing_leads import _source_category
+
+    lead = db.query(SourcingLead).filter(SourcingLead.id == lead_id).first()
+    if not lead:
+        raise HTTPException(404, "Lead not found")
+
+    evidence = (
+        db.query(LeadEvidence)
+        .filter(LeadEvidence.lead_id == lead.id)
+        .order_by(LeadEvidence.confidence_impact.desc().nullslast())
+        .all()
+    )
+
+    evidence_by_category = {}
+    for ev in evidence:
+        cat = _source_category(ev.source_type)
+        evidence_by_category.setdefault(cat, []).append(ev)
+
+    category_labels = {
+        "api": "API",
+        "marketplace": "Marketplace",
+        "salesforce_history": "Salesforce History",
+        "avail_history": "Avail History",
+        "web_ai": "Web / AI",
+        "safety_review": "Safety Review",
+        "buyer_feedback": "Buyer Feedback",
+    }
+
+    requirement = db.query(Requirement).filter(Requirement.id == lead.requirement_id).first()
+    vendor_card = lead.vendor_card
+
+    best_sighting = (
+        db.query(Sighting)
+        .filter(
+            Sighting.requirement_id == lead.requirement_id,
+            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
+        )
+        .order_by(Sighting.created_at.desc().nullslast())
+        .first()
+    )
+
+    ctx = _base_ctx(request, user, "requisitions")
+    ctx.update(
+        {
+            "lead": lead,
+            "evidence": evidence,
+            "evidence_by_category": evidence_by_category,
+            "category_labels": category_labels,
+            "requirement": requirement,
+            "vendor_card": vendor_card,
+            "best_sighting": best_sighting,
+        }
+    )
+    return template_response("htmx/partials/sourcing/lead_panel.html", ctx)
