+++ app/routers/htmx_views.py
@@ -12,8 +12,7 @@ import asyncio
 import html as html_mod
 import json
 import os
-import time
-from datetime import date, datetime, timedelta, timezone
+from datetime import date, datetime, timezone
 from urllib.parse import quote
 
 from fastapi import APIRouter, Depends, Form, HTTPException, Query, Request
@@ -21,16 +20,12 @@ from fastapi.responses import HTMLResponse
 from loguru import logger
 from sqlalchemy import case, desc, or_, select
 from sqlalchemy import func as sqlfunc
-from sqlalchemy.orm import Session, joinedload, selectinload
+from sqlalchemy.orm import Session, joinedload
 
 from ..constants import (
-    RESTRICTED_ROLES,
     AccessKey,
     ActivityType,
     ApiSourceStatus,
-    AttributionStatus,
-    BuyPlanStatus,
-    ContactStatus,
     OfferStatus,
     ProactiveMatchStatus,
     ProspectAccountStatus,
@@ -41,13 +36,10 @@ from ..constants import (
     TicketSource,
     TicketStatus,
     UserRole,
-    VendorResponseStatus,
 )
 from ..database import get_db
 from ..dependencies import (
-    can_approve_buy_plans,
     can_manage_account,
-    get_buyplan_for_user,
     get_quote_for_user,
     get_req_for_user,
     get_user,
@@ -56,14 +48,12 @@ from ..dependencies import (
     require_access,
     require_admin,
     require_buyer,
-    require_buyplan_approver,
     require_requisition_access,
     require_user,
     user_has_access,
 )
 from ..models import (
     ApiSource,
-    BuyPlan,
     BuyPlanLine,
     Company,
     CustomerSite,
@@ -78,7 +68,6 @@ from ..models import (
     SourcingLead,
     User,
     VendorCard,
-    VerificationGroupMember,
 )
 from ..models.faceted_search import CommoditySpecSchema
 from ..models.prospect_account import ProspectAccount
@@ -113,13 +102,13 @@ from ..services.quote_send import (
 from ..services.sighting_aggregation import get_vendor_tier_map
 from ..services.sighting_ingest import sighting_from_row
 from ..services.status_machine import require_valid_transition
-from ..services.vendor_unavailability import apply_to_fresh_sightings, maybe_release_on_offer
+from ..services.vendor_unavailability import apply_to_fresh_sightings
 from ..template_env import page_response, template_response, templates
 from ..utils.search_builder import SearchBuilder
 from ..utils.sql_helpers import escape_like
 from ._lookup_helpers import get_requisition_or_404, get_vendor_card_or_404
 from .auth import _password_login_enabled
-from .htmx._shared import _base_ctx, _parse_date_safe, _vite_assets
+from .htmx._shared import _base_ctx, _is_ops_member, _safe_int, _vite_assets
 from .htmx.companies import company_tab
 from .htmx.requisitions import requisition_tab, requisitions_list_partial
 from .htmx.vendors import vendor_tab
@@ -458,20 +447,6 @@ async def v2_page(request: Request, db: Session = Depends(get_db)):
     return page_response(ctx)
 
 
-@router.get("/v2/buy-plans", response_class=HTMLResponse)
-async def buy_plans_legacy_redirect(request: Request):
-    """302 the legacy Buy Plans URL to the renamed Approvals module (query preserved).
-
-    The hub was renamed Buy Plans → Approvals (SP-1); old bookmarks / pushed lens URLs
-    keep working via this redirect. Detail URLs (/v2/buy-plans/{id}) are unchanged and
-    still served directly by ``v2_page``.
-    """
-    from fastapi.responses import RedirectResponse
-
-    qs = request.url.query
-    return RedirectResponse(f"/v2/approvals?{qs}" if qs else "/v2/approvals", status_code=302)
-
-
 # ── Global search ──────────────────────────────────────────────────────
 
 
@@ -588,264 +563,6 @@ async def customer_activity_digest(
     ctx["digest"] = digest
     ctx["refresh_url"] = f"/v2/partials/customers/{company_id}/activity-digest"
     return template_response("htmx/partials/shared/activity_digest_card.html", ctx)
-
-
-# ── Column Prefs Save Endpoints ──────────────────────────────────────────────
-
-
-# ── AI Parsing in Requisition Offers (Phase 3B) ───────────────────────
-
-
-@router.get("/v2/partials/requisitions/{req_id}/parse-email-form", response_class=HTMLResponse)
-async def parse_email_form(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return the parse-email paste form."""
-    req = get_requisition_or_404(db, req_id)
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    return template_response("htmx/partials/requisitions/tabs/parse_email_form.html", ctx)
-
-
-@router.get("/v2/partials/requisitions/{req_id}/paste-offer-form", response_class=HTMLResponse)
-async def paste_offer_form(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return the paste-offer freeform form."""
-    req = get_requisition_or_404(db, req_id)
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    return template_response("htmx/partials/requisitions/tabs/paste_offer_form.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/parse-email", response_class=HTMLResponse)
-async def parse_email_action(
-    request: Request,
-    req_id: int,
-    email_body: str = Form(""),
-    email_subject: str = Form(""),
-    vendor_name: str = Form(""),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Parse vendor email and return editable offer cards."""
-    req = get_requisition_or_404(db, req_id)
-    require_requisition_access(db, req_id, user)
-
-    if not email_body.strip():
-        return HTMLResponse(
-            '<div class="p-4 text-center text-sm text-amber-600 bg-amber-50 rounded-lg border border-amber-200">'
-            "Please paste the email body to parse.</div>"
-        )
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    ctx["vendor_name"] = vendor_name
-
-    try:
-        from app.services.ai_email_parser import parse_email
-
-        result = await parse_email(
-            email_body=email_body,
-            email_subject=email_subject,
-            vendor_name=vendor_name,
-        )
-
-        if not result:
-            ctx["quotes"] = []
-            ctx["overall_confidence"] = 0
-            ctx["email_type"] = "unclear"
-        else:
-            ctx["quotes"] = result.get("quotes", [])
-            ctx["overall_confidence"] = result.get("overall_confidence", 0)
-            ctx["email_type"] = result.get("email_type", "unclear")
-
-    except Exception as exc:
-        logger.error(f"Parse email error for req {req_id}: {exc}")
-        return HTMLResponse(
-            '<div class="p-4 text-center text-sm text-rose-600 bg-rose-50 rounded-lg border border-rose-200">'
-            f"Parse failed: {exc}</div>"
-        )
-
-    return template_response("htmx/partials/requisitions/tabs/parsed_email_results.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/parse-offer", response_class=HTMLResponse)
-async def parse_offer_action(
-    request: Request,
-    req_id: int,
-    raw_text: str = Form(""),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Parse freeform vendor text and return editable offer cards."""
-    req = get_requisition_or_404(db, req_id)
-    require_requisition_access(db, req_id, user)
-
-    if not raw_text.strip():
-        return HTMLResponse(
-            '<div class="p-4 text-center text-sm text-amber-600 bg-amber-50 rounded-lg border border-amber-200">'
-            "Please paste vendor text to parse.</div>"
-        )
-
-    # Build RFQ context for better matching
-    reqs = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-    rfq_context = [{"mpn": r.primary_mpn, "qty": r.target_qty or 1} for r in reqs if r.primary_mpn]
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-
-    try:
-        from app.services.freeform_parser_service import parse_freeform_offer
-
-        result = await parse_freeform_offer(raw_text, rfq_context)
-        if not result:
-            ctx["offers"] = []
-        else:
-            ctx["offers"] = result.get("offers", [])
-    except Exception as exc:
-        logger.error(f"Parse offer error for req {req_id}: {exc}")
-        return HTMLResponse(
-            '<div class="p-4 text-center text-sm text-rose-600 bg-rose-50 rounded-lg border border-rose-200">'
-            f"Parse failed: {exc}</div>"
-        )
-
-    return template_response("htmx/partials/requisitions/tabs/parsed_offer_results.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/save-parsed-offers", response_class=HTMLResponse)
-async def save_parsed_offers(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Save user-edited parsed offers to the requisition."""
-    req = get_requisition_or_404(db, req_id)
-    require_requisition_access(db, req_id, user)
-
-    form = await request.form()
-    vendor_name = form.get("vendor_name", "")
-
-    # Collect offers from form fields (offers[0].mpn, offers[0].qty_available, etc.)
-    offers_data: list[dict] = []
-    idx = 0
-    while True:
-        mpn = form.get(f"offers[{idx}].mpn")
-        if mpn is None:
-            # Also check vendor_name field for freeform offers
-            vn = form.get(f"offers[{idx}].vendor_name")
-            if vn is None:
-                break
-        offer = {
-            "vendor_name": form.get(f"offers[{idx}].vendor_name", vendor_name),
-            "mpn": form.get(f"offers[{idx}].mpn", ""),
-            "manufacturer": form.get(f"offers[{idx}].manufacturer"),
-            "qty_available": _safe_int(form.get(f"offers[{idx}].qty_available")),
-            "unit_price": _safe_float(form.get(f"offers[{idx}].unit_price")),
-            "lead_time": form.get(f"offers[{idx}].lead_time"),
-            "date_code": form.get(f"offers[{idx}].date_code"),
-            "condition": form.get(f"offers[{idx}].condition", "new"),
-            "moq": _safe_int(form.get(f"offers[{idx}].moq")),
-            "notes": form.get(f"offers[{idx}].notes"),
-        }
-        offers_data.append(offer)
-        idx += 1
-
-    if not offers_data:
-        return HTMLResponse(
-            '<div class="p-4 text-center text-sm text-amber-600 bg-amber-50 rounded-lg border border-amber-200">'
-            "No offers to save.</div>"
-        )
-
-    # Match MPNs to requirements
-    reqs = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-    from app.vendor_utils import normalize_vendor_name
-
-    saved_count = 0
-    for o in offers_data:
-        if not o["mpn"]:
-            continue
-
-        # Find matching requirement
-        req_match_id = None
-        mpn_lower = (o["mpn"] or "").strip().lower()
-        for r in reqs:
-            if r.primary_mpn and r.primary_mpn.strip().lower() == mpn_lower:
-                req_match_id = r.id
-                break
-
-        # Resolve vendor card
-        vn = o.get("vendor_name") or vendor_name or "Unknown"
-        norm_name = normalize_vendor_name(vn)
-        card = db.query(VendorCard).filter(VendorCard.normalized_name == norm_name).first()
-        if not card:
-            card = VendorCard(
-                normalized_name=norm_name,
-                display_name=vn,
-                emails=[],
-                phones=[],
-            )
-            db.add(card)
-            db.flush()
-
-        offer = Offer(
-            requisition_id=req_id,
-            requirement_id=req_match_id,
-            vendor_card_id=card.id,
-            vendor_name=card.display_name,
-            vendor_name_normalized=card.normalized_name,
-            mpn=o["mpn"],
-            manufacturer=o.get("manufacturer"),
-            qty_available=o.get("qty_available"),
-            unit_price=o.get("unit_price"),
-            lead_time=o.get("lead_time"),
-            date_code=o.get("date_code"),
-            condition=o.get("condition") or "new",
-            moq=o.get("moq"),
-            notes=o.get("notes"),
-            source="ai_parsed",
-            entered_by_id=user.id,
-            status=OfferStatus.ACTIVE,
-        )
-        db.add(offer)
-        # Offer hook: the user reviewed and saved this parse ACTIVE — user-initiated
-        # proof of availability, release the vendor's matching active records.
-        maybe_release_on_offer(db, req_match_id, offer.vendor_name, user)
-        saved_count += 1
-
-    db.commit()
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    ctx["saved_count"] = saved_count
-    return template_response("htmx/partials/requisitions/tabs/parse_save_success.html", ctx)
-
-
-def _safe_int(val) -> int | None:
-    """Safely convert form value to int."""
-    if not val:
-        return None
-    try:
-        return int(val)
-    except (ValueError, TypeError):
-        return None
-
-
-def _safe_float(val) -> float | None:
-    """Safely convert form value to float."""
-    if not val:
-        return None
-    try:
-        return float(val)
-    except (ValueError, TypeError):
-        return None
 
 
 @router.post("/v2/partials/requisitions/bulk/{action}", response_class=HTMLResponse)
@@ -1197,1323 +914,6 @@ async def requisition_row_action(
 
     response.headers["HX-Trigger"] = json.dumps({"showToast": {"message": msg}})
     return response
-
-
-@router.post("/v2/partials/requisitions/{req_id}/create-quote", response_class=HTMLResponse)
-async def create_quote_from_offers(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Create a new quote from selected offer IDs.
-
-    Returns quote detail partial.
-    """
-    form = await request.form()
-    offer_ids_raw = form.getlist("offer_ids")
-    try:
-        offer_ids = [int(x) for x in offer_ids_raw if x]
-    except (ValueError, TypeError):
-        raise HTTPException(400, "offer_ids must be integers")
-
-    if not offer_ids:
-        raise HTTPException(400, "No offers selected")
-
-    req = get_requisition_or_404(db, req_id)
-    require_requisition_access(db, req_id, user)
-
-    offers = db.query(Offer).filter(Offer.id.in_(offer_ids), Offer.requisition_id == req_id).all()
-    if not offers:
-        raise HTTPException(404, "No matching offers found")
-
-    # Build line items from offers
-
-    quote_number = f"Q-{req_id}-{db.query(Quote).filter(Quote.requisition_id == req_id).count() + 1}"
-    quote = Quote(
-        requisition_id=req_id,
-        quote_number=quote_number,
-        status=QuoteStatus.DRAFT,
-        created_by_id=user.id,
-        customer_site_id=req.customer_site_id,
-    )
-    db.add(quote)
-    db.flush()
-
-    subtotal = 0.0
-    total_cost = 0.0
-    for o in offers:
-        sell_price = float(o.unit_price or 0)
-        cost_price = sell_price  # Default cost = sell, buyer adjusts
-        qty = o.qty_available or 1
-        margin_pct = 0.0
-
-        line = QuoteLine(
-            quote_id=quote.id,
-            offer_id=o.id,
-            mpn=o.mpn or "",
-            manufacturer=o.manufacturer or "",
-            qty=qty,
-            cost_price=cost_price,
-            sell_price=sell_price,
-            margin_pct=margin_pct,
-        )
-        db.add(line)
-        subtotal += sell_price * qty
-        total_cost += cost_price * qty
-
-    quote.subtotal = subtotal
-    quote.total_cost = total_cost
-    quote.total_margin_pct = ((subtotal - total_cost) / subtotal * 100) if subtotal else 0
-    db.commit()
-    db.refresh(quote)
-
-    logger.info("Created quote {} from {} offers by {}", quote.quote_number, len(offers), user.email)
-
-    # Return the quote detail page
-    lines = db.query(QuoteLine).filter(QuoteLine.quote_id == quote.id).all()
-    ctx = _base_ctx(request, user, "quotes")
-    ctx["quote"] = quote
-    ctx["lines"] = lines
-    ctx["offers"] = offers
-    return template_response("htmx/partials/quotes/detail.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/offers/{offer_id}/review", response_class=HTMLResponse)
-async def review_offer(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Approve or reject an offer.
-
-    Returns refreshed offers tab.
-    """
-    form = await request.form()
-    action = form.get("action", "")
-
-    if action not in ("approve", "reject"):
-        raise HTTPException(400, "Invalid action")
-
-    require_requisition_access(db, req_id, user)
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-
-    old_status = offer.status
-
-    if action == "approve":
-        require_valid_transition("offer", offer.status, OfferStatus.APPROVED)
-        offer.status = OfferStatus.APPROVED
-        offer.approved_by_id = user.id
-
-        offer.approved_at = datetime.now(timezone.utc)
-        # Offer hook: user approval of a pending offer is user-initiated proof of
-        # availability — release the vendor's matching active unavailability records.
-        maybe_release_on_offer(db, offer.requirement_id, offer.vendor_name, user)
-    else:
-        require_valid_transition("offer", offer.status, OfferStatus.REJECTED)
-        offer.status = OfferStatus.REJECTED
-
-    _log_activity(
-        db,
-        activity_type=ActivityType.OFFER_STATUS_CHANGED,
-        requisition_id=offer.requisition_id,
-        user_id=user.id,
-        vendor_card_id=offer.vendor_card_id,
-        description=f"Offer {offer.vendor_name} status: {old_status} → {offer.status}",
-        details={
-            "offer_id": offer.id,
-            "old_status": str(old_status),
-            "new_status": str(offer.status),
-        },
-    )
-
-    db.commit()
-    logger.info("Offer {} {} by {}", offer_id, action, user.email)
-
-    # Return refreshed offers tab
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-@router.get("/v2/partials/requisitions/{req_id}/add-offer-form", response_class=HTMLResponse)
-async def add_offer_form(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return the manual offer entry form."""
-    req = get_requisition_or_404(db, req_id)
-    requirements = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    ctx["requirements"] = requirements
-    return template_response("htmx/partials/requisitions/add_offer_form.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/add-offer", response_class=HTMLResponse)
-async def add_offer(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Create a manual offer and return refreshed offers tab."""
-    get_requisition_or_404(db, req_id)  # validates existence
-    require_requisition_access(db, req_id, user)
-
-    form = await request.form()
-    vendor_name = (form.get("vendor_name") or "").strip()
-    mpn = (form.get("mpn") or "").strip()
-    if not vendor_name or not mpn:
-        return HTMLResponse(
-            '<div class="p-3 text-sm text-rose-600 bg-rose-50 rounded mb-4">Vendor name and MPN are required.</div>',
-            status_code=400,
-        )
-
-    from datetime import date as date_type
-
-    from ..services.offer_qualification import (
-        apply_qualification,
-        normalize_offer_condition,
-    )
-    from ..utils.normalization import normalize_mpn_key
-    from ..vendor_utils import normalize_vendor_name
-
-    offer = Offer(
-        requisition_id=req_id,
-        vendor_name=vendor_name,
-        vendor_name_normalized=normalize_vendor_name(vendor_name),
-        mpn=mpn,
-        # Canonical dedup key (dash-stripped) so the part-centric offers query
-        # matches consistently with create_offer's normalize_mpn_key.
-        normalized_mpn=normalize_mpn_key(mpn),
-        qty_available=_safe_int(form.get("qty_available")),
-        unit_price=_safe_float(form.get("unit_price")),
-        lead_time=form.get("lead_time") or None,
-        date_code=form.get("date_code") or None,
-        condition=normalize_offer_condition(form.get("condition")) or form.get("condition") or None,
-        moq=_safe_int(form.get("moq")),
-        manufacturer=form.get("manufacturer") or None,
-        spq=_safe_int(form.get("spq")),
-        packaging=form.get("packaging") or None,
-        firmware=form.get("firmware") or None,
-        hardware_code=form.get("hardware_code") or None,
-        warranty=form.get("warranty") or None,
-        country_of_origin=form.get("country_of_origin") or None,
-        valid_until=_parse_date_safe(form.get("valid_until"), date_type),
-        notes=form.get("notes") or None,
-        requirement_id=_safe_int(form.get("requirement_id")),
-        source="manual",
-        status=OfferStatus.ACTIVE,
-        entered_by_id=user.id,
-        created_at=datetime.now(timezone.utc),
-    )
-    _qkeys = (
-        "usage",
-        "refurbished_by",
-        "refurb_process",
-        "cert_doc",
-        "part_condition",
-        "provenance_story",
-        "terms",
-        "lead_time_reason",
-    )
-    qual = {k: (form.get(k) or None) for k in _qkeys}
-    qual["requests"] = []
-    qual["schema"] = 1  # forward-version the qualification blob (spec §3.1)
-    offer.qualification = qual if any(qual[k] for k in _qkeys) else None
-    apply_qualification(offer)  # non-raising: composes note + sets status
-    db.add(offer)
-    db.flush()  # offer.id populated; activity row + offer committed together below
-    # Offer hook: a manually entered offer is user-initiated proof of availability —
-    # release the vendor's matching active unavailability records.
-    maybe_release_on_offer(db, offer.requirement_id, offer.vendor_name, user)
-    logger.info("Manual offer created: {} on req {} by {}", mpn, req_id, user.email)
-
-    _log_activity(
-        db,
-        activity_type=ActivityType.OFFER_CREATED,
-        requisition_id=offer.requisition_id,
-        requirement_id=offer.requirement_id,
-        user_id=user.id,
-        vendor_card_id=offer.vendor_card_id,
-        description=f"Offer added: {offer.vendor_name} — {offer.mpn}",
-        details={"offer_id": offer.id, "source": offer.source},
-    )
-    db.commit()
-
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/offers/{offer_id}/reconfirm", response_class=HTMLResponse)
-async def reconfirm_offer(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Reconfirm an offer — resets TTL and increments reconfirm count."""
-    require_requisition_access(db, req_id, user)
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-
-    now = datetime.now(timezone.utc)
-    offer.reconfirmed_at = now
-    offer.reconfirm_count = (offer.reconfirm_count or 0) + 1
-    offer.expires_at = now + timedelta(days=14)
-    offer.attribution_status = AttributionStatus.ACTIVE
-    offer.is_stale = False
-    offer.updated_at = now
-    offer.updated_by_id = user.id
-    db.commit()
-    logger.info("Offer {} reconfirmed (count={}) by {}", offer_id, offer.reconfirm_count, user.email)
-
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-# ── Sprint 2: Offer Management Completion ─────────────────────────────
-
-
-@router.get("/v2/partials/requisitions/{req_id}/offers/{offer_id}/edit-form", response_class=HTMLResponse)
-async def edit_offer_form(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return inline edit form for an existing offer."""
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    requirements = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-    return template_response(
-        "htmx/partials/requisitions/edit_offer_form.html",
-        {"request": request, "offer": offer, "req_id": req_id, "requirements": requirements},
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/offers/{offer_id}/edit", response_class=HTMLResponse)
-async def edit_offer(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Save edits to an offer and return refreshed offers tab."""
-    from ..models.intelligence import ChangeLog
-
-    require_requisition_access(db, req_id, user)
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-
-    form = await request.form()
-    trackable = [
-        "vendor_name",
-        "qty_available",
-        "unit_price",
-        "lead_time",
-        "condition",
-        "date_code",
-        "moq",
-        "notes",
-        "manufacturer",
-        "spq",
-        "packaging",
-        "firmware",
-        "hardware_code",
-        "warranty",
-        "country_of_origin",
-        "valid_until",
-    ]
-    now = datetime.now(timezone.utc)
-
-    for field in trackable:
-        new_val = form.get(field, "").strip()
-        old_val = str(getattr(offer, field) or "")
-        if new_val != old_val and new_val:
-            if field in ("qty_available", "moq", "spq"):
-                try:
-                    setattr(offer, field, int(new_val))
-                except ValueError:
-                    continue
-            elif field == "unit_price":
-                try:
-                    setattr(offer, field, float(new_val))
-                except ValueError:
-                    continue
-            elif field == "valid_until":
-                from datetime import date as date_type
-
-                try:
-                    setattr(offer, field, date_type.fromisoformat(new_val) if new_val else None)
-                except ValueError:
-                    continue
-            else:
-                setattr(offer, field, new_val)
-            db.add(
-                ChangeLog(
-                    entity_type="offer",
-                    entity_id=offer_id,
-                    user_id=user.id,
-                    field_name=field,
-                    old_value=old_val,
-                    new_value=new_val,
-                )
-            )
-
-    req_id_val = form.get("requirement_id", "")
-    if req_id_val:
-        offer.requirement_id = int(req_id_val) if req_id_val.isdigit() else None
-
-    from ..services.offer_qualification import (
-        apply_qualification,
-        normalize_offer_condition,
-    )
-
-    _qkeys = (
-        "usage",
-        "refurbished_by",
-        "refurb_process",
-        "cert_doc",
-        "part_condition",
-        "provenance_story",
-        "terms",
-        "lead_time_reason",
-    )
-    submitted_qual = {k: (form.get(k) or None) for k in _qkeys}
-    if any(submitted_qual.values()):
-        merged = dict(offer.qualification or {})
-        merged.update(submitted_qual)
-        merged.setdefault("requests", [])
-        merged["schema"] = 1  # forward-version the qualification blob (spec §3.1)
-        offer.qualification = merged
-    cond_raw = form.get("condition", "").strip()
-    if cond_raw:
-        offer.condition = normalize_offer_condition(cond_raw) or cond_raw
-
-    apply_qualification(offer)  # non-raising: composes note + sets status
-    offer.updated_at = now
-    offer.updated_by_id = user.id
-    db.commit()
-    logger.info("Offer {} edited by {}", offer_id, user.email)
-
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-@router.delete("/v2/partials/requisitions/{req_id}/offers/{offer_id}", response_class=HTMLResponse)
-async def delete_offer_htmx(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Delete an offer and return refreshed offers tab."""
-    require_requisition_access(db, req_id, user)
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    db.delete(offer)
-    db.commit()
-    logger.info("Offer {} deleted by {}", offer_id, user.email)
-
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/offers/{offer_id}/mark-sold", response_class=HTMLResponse)
-async def mark_offer_sold_htmx(
-    request: Request,
-    req_id: int,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Mark an offer as sold and return refreshed offers tab."""
-    from ..models.intelligence import ChangeLog
-
-    require_requisition_access(db, req_id, user)
-    offer = db.query(Offer).filter(Offer.id == offer_id, Offer.requisition_id == req_id).first()
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    if offer.status == OfferStatus.SOLD:
-        return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-    old_status = offer.status
-    require_valid_transition("offer", offer.status, OfferStatus.SOLD)
-    offer.status = OfferStatus.SOLD
-    offer.updated_at = datetime.now(timezone.utc)
-    offer.updated_by_id = user.id
-    db.add(
-        ChangeLog(
-            entity_type="offer",
-            entity_id=offer_id,
-            user_id=user.id,
-            field_name="status",
-            old_value=old_status,
-            new_value="sold",
-        )
-    )
-
-    _log_activity(
-        db,
-        activity_type=ActivityType.OFFER_STATUS_CHANGED,
-        requisition_id=offer.requisition_id,
-        user_id=user.id,
-        vendor_card_id=offer.vendor_card_id,
-        description=f"Offer {offer.vendor_name} status: {old_status} → {offer.status}",
-        details={
-            "offer_id": offer.id,
-            "old_status": str(old_status),
-            "new_status": str(offer.status),
-        },
-    )
-
-    db.commit()
-    logger.info("Offer {} marked sold by {}", offer_id, user.email)
-
-    return await requisition_tab(request=request, req_id=req_id, tab="offers", user=user, db=db)
-
-
-@router.get("/v2/partials/offers/review-queue", response_class=HTMLResponse)
-async def offer_review_queue(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Render the offer review queue page — medium-confidence AI-parsed offers."""
-    offers = (
-        db.query(Offer)
-        .filter(Offer.status == OfferStatus.PENDING_REVIEW)
-        .order_by(Offer.created_at.desc())
-        .limit(100)
-        .all()
-    )
-    return template_response(
-        "htmx/partials/offers/review_queue.html",
-        {"request": request, "offers": offers, "user": user},
-    )
-
-
-@router.post("/v2/partials/offers/{offer_id}/promote", response_class=HTMLResponse)
-async def promote_offer_htmx(
-    request: Request,
-    offer_id: int,
-    user: User = Depends(require_access(AccessKey.APPROVE_OFFERS)),
-    db: Session = Depends(get_db),
-):
-    """Promote a pending_review offer to active and return refreshed queue."""
-    offer = db.get(Offer, offer_id)
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    require_requisition_access(db, offer.requisition_id, user, owner_id=offer.entered_by_id, label="Offer")
-    if offer.status != OfferStatus.PENDING_REVIEW:
-        raise HTTPException(400, "Only pending_review offers can be promoted")
-
-    old_status = offer.status
-    require_valid_transition("offer", offer.status, OfferStatus.ACTIVE)
-    offer.status = OfferStatus.ACTIVE
-    offer.approved_by_id = user.id
-    offer.approved_at = datetime.now(timezone.utc)
-    offer.updated_at = datetime.now(timezone.utc)
-    offer.updated_by_id = user.id
-
-    # Offer hook: user approval of a pending offer is user-initiated proof of
-    # availability — release the vendor's matching active unavailability records.
-    maybe_release_on_offer(db, offer.requirement_id, offer.vendor_name, user)
-
-    _log_activity(
-        db,
-        activity_type=ActivityType.OFFER_STATUS_CHANGED,
-        requisition_id=offer.requisition_id,
-        user_id=user.id,
-        vendor_card_id=offer.vendor_card_id,
-        description=f"Offer {offer.vendor_name} status: {old_status} → {offer.status}",
-        details={
-            "offer_id": offer.id,
-            "old_status": str(old_status),
-            "new_status": str(offer.status),
-        },
-    )
-
-    db.commit()
-    logger.info("Offer {} promoted by {}", offer_id, user.email)
-
-    return await offer_review_queue(request=request, user=user, db=db)
-
-
-@router.post("/v2/partials/offers/{offer_id}/reject", response_class=HTMLResponse)
-async def reject_offer_htmx(
-    request: Request,
-    offer_id: int,
-    user: User = Depends(require_access(AccessKey.APPROVE_OFFERS)),
-    db: Session = Depends(get_db),
-):
-    """Reject a pending_review offer and return refreshed queue."""
-    offer = db.get(Offer, offer_id)
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    require_requisition_access(db, offer.requisition_id, user, owner_id=offer.entered_by_id, label="Offer")
-    if offer.status != OfferStatus.PENDING_REVIEW:
-        raise HTTPException(400, "Only pending_review offers can be rejected")
-
-    old_status = offer.status
-    require_valid_transition("offer", offer.status, OfferStatus.REJECTED)
-    offer.status = OfferStatus.REJECTED
-    offer.updated_at = datetime.now(timezone.utc)
-    offer.updated_by_id = user.id
-
-    _log_activity(
-        db,
-        activity_type=ActivityType.OFFER_STATUS_CHANGED,
-        requisition_id=offer.requisition_id,
-        user_id=user.id,
-        vendor_card_id=offer.vendor_card_id,
-        description=f"Offer {offer.vendor_name} status: {old_status} → {offer.status}",
-        details={
-            "offer_id": offer.id,
-            "old_status": str(old_status),
-            "new_status": str(offer.status),
-        },
-    )
-
-    db.commit()
-    logger.info("Offer {} rejected by {}", offer_id, user.email)
-
-    return await offer_review_queue(request=request, user=user, db=db)
-
-
-@router.get("/v2/partials/offers/{offer_id}/changelog", response_class=HTMLResponse)
-async def offer_changelog(
-    request: Request,
-    offer_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Render change history for an offer."""
-    from ..models.intelligence import ChangeLog
-
-    offer = db.get(Offer, offer_id)
-    if not offer:
-        raise HTTPException(404, "Offer not found")
-    rows = (
-        db.query(ChangeLog)
-        .filter(ChangeLog.entity_type == "offer", ChangeLog.entity_id == offer_id)
-        .options(joinedload(ChangeLog.user))
-        .order_by(ChangeLog.created_at.desc())
-        .limit(50)
-        .all()
-    )
-    return template_response(
-        "htmx/partials/offers/changelog.html",
-        {"request": request, "offer": offer, "changes": rows},
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/log-activity", response_class=HTMLResponse)
-async def log_activity(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Log a manual activity (note/call/email) for a requisition."""
-    from ..models.intelligence import ActivityLog
-
-    get_requisition_or_404(db, req_id)  # validates existence
-    require_requisition_access(db, req_id, user)
-
-    form = await request.form()
-    activity_type = form.get("activity_type", "note")
-    channel_map = {"note": "note", "phone_call": "phone", "email_sent": "email"}
-
-    log = ActivityLog(
-        user_id=user.id,
-        requisition_id=req_id,
-        activity_type=activity_type,
-        channel=channel_map.get(activity_type, "note"),
-        contact_name=form.get("vendor_name", ""),
-        contact_phone=form.get("contact_phone", ""),
-        contact_email=form.get("contact_email", ""),
-        notes=form.get("notes", ""),
-    )
-    db.add(log)
-    db.commit()
-    db.refresh(log)
-    logger.info("Activity logged for req {} by {}: {}", req_id, user.email, activity_type)
-
-    # Return refreshed activity tab
-    return await requisition_tab(request=request, req_id=req_id, tab="activity", user=user, db=db)
-
-
-@router.get("/v2/partials/requisitions/{req_id}/rfq-compose", response_class=HTMLResponse)
-async def rfq_compose(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Render the RFQ compose form for a requisition."""
-    from ..models.offers import Contact as RfqContact
-
-    req = get_requisition_or_404(db, req_id)
-
-    parts = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-
-    # Get unique vendors from sightings for this requisition's parts
-    part_ids = [p.id for p in parts]
-    vendors = []
-    if part_ids:
-        # Get distinct vendor names from sightings, then match to VendorCard
-        vendor_names = (
-            db.query(Sighting.vendor_name_normalized)
-            .filter(Sighting.requirement_id.in_(part_ids), Sighting.vendor_name_normalized.isnot(None))
-            .distinct()
-            .all()
-        )
-        norm_names = [n[0] for n in vendor_names if n[0]]
-        vendor_rows = (
-            (
-                db.query(VendorCard)
-                .options(selectinload(VendorCard.vendor_contacts))
-                .filter(VendorCard.normalized_name.in_(norm_names))
-                .limit(50)
-                .all()
-            )
-            if norm_names
-            else []
-        )
-        # Check which vendors already have RFQs sent
-        sent_vendor_names = set()
-        existing_contacts = db.query(RfqContact).filter(RfqContact.requisition_id == req_id).all()
-        for c in existing_contacts:
-            if c.vendor_name_normalized:
-                sent_vendor_names.add(c.vendor_name_normalized)
-
-        for v in vendor_rows:
-            # Get contacts for this vendor
-            v_contacts = v.vendor_contacts[:5]  # Already eagerly loaded
-            vendors.append(
-                {
-                    "id": v.id,
-                    "display_name": v.display_name,
-                    "normalized_name": v.normalized_name,
-                    "domain": v.domain,
-                    "contacts": v_contacts,
-                    "already_asked": v.normalized_name in sent_vendor_names,
-                    "emails": [c.email for c in v_contacts if c.email],
-                }
-            )
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    ctx["parts"] = parts
-    ctx["vendors"] = vendors
-    return template_response("htmx/partials/requisitions/rfq_compose.html", ctx)
-
-
-@router.post("/v2/partials/requisitions/{req_id}/ai-cleanup-email", response_class=HTMLResponse)
-async def ai_cleanup_email(
-    request: Request,
-    req_id: int,
-    body: str = Form(""),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Clean up user-written email — fix grammar, tone, and formatting."""
-    get_requisition_or_404(db, req_id)  # validates existence
-    require_requisition_access(db, req_id, user)
-
-    user_text = body.strip()
-    if not user_text:
-        return HTMLResponse('<p class="text-xs text-amber-600 mt-1">Write your email first, then click Clean Up.</p>')
-
-    try:
-        from app.utils.claude_client import claude_text
-
-        result = await claude_text(
-            prompt=(
-                f"Clean up this RFQ email: fix grammar, spelling, punctuation. "
-                f"Improve clarity and professional tone. Keep it concise. "
-                f"Do NOT add information the user didn't include. "
-                f"Do NOT change the meaning or add new requests. "
-                f"Return ONLY the cleaned-up email text, nothing else.\n\n"
-                f"---\n{user_text}\n---"
-            ),
-            system="You are an email editor for a professional electronic components buyer.",
-            model_tier="fast",
-            max_tokens=1000,
-        )
-        cleaned = result.strip() if result else user_text
-    except Exception as exc:
-        logger.error("AI cleanup error for req {}: {}", req_id, exc)
-        cleaned = user_text
-
-    # Return a script that replaces the textarea content with the cleaned text
-    escaped = cleaned.replace("\\", "\\\\").replace("`", "\\`").replace("$", "\\$").replace("</", "<\\/")
-    return HTMLResponse(
-        f'<script>document.getElementById("rfq-body-textarea").value = `{escaped}`;</script>'
-        '<p class="text-xs text-green-600 mt-1">Email cleaned up. Review and edit as needed.</p>'
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/ai-rephrase-email", response_class=HTMLResponse)
-async def ai_rephrase_email(
-    request: Request,
-    req_id: int,
-    body: str = Form(""),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Rephrase the RFQ email so each send reads uniquely, keeping all parts intact."""
-    get_requisition_or_404(db, req_id)  # validates existence
-    require_requisition_access(db, req_id, user)
-
-    user_text = body.strip()
-    if not user_text:
-        return HTMLResponse(
-            '<p class="text-xs text-amber-600 mt-1">Write your email first, then click AI Rephrase.</p>'
-        )
-
-    from app.services.email_drafting import draft_email
-
-    result = await draft_email("rfq_rephrase", {"body": user_text})
-    rephrased = (result or {}).get("body") or user_text
-
-    # Return a script that replaces the textarea content with the rephrased text.
-    escaped = rephrased.replace("\\", "\\\\").replace("`", "\\`").replace("$", "\\$").replace("</", "<\\/")
-    return HTMLResponse(
-        f'<script>document.getElementById("rfq-body-textarea").value = `{escaped}`;</script>'
-        '<p class="text-xs text-green-600 mt-1">Rephrased. Review and edit as needed.</p>'
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/rfq-send", response_class=HTMLResponse)
-async def rfq_send(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_access(AccessKey.SEND_RFQ)),
-    db: Session = Depends(get_db),
-):
-    """Send RFQs via Graph API, falling back to DB-only in test mode."""
-    import os
-
-    from ..models.offers import Contact as RfqContact
-
-    req = get_requisition_or_404(db, req_id)
-    require_requisition_access(db, req_id, user)
-
-    form = await request.form()
-    vendor_names = form.getlist("vendor_names")
-    vendor_emails = form.getlist("vendor_emails")
-    subject = form.get("subject", f"RFQ - {req.name}")
-    body = form.get("body", "")
-    parts_text = form.get("parts_summary", "")
-
-    if not vendor_names:
-        raise HTTPException(400, "No vendors selected")
-
-    # Try to get a fresh Graph API token for real email send
-    token = None
-    is_testing = os.environ.get("TESTING") == "1"
-    if not is_testing:
-        try:
-            from ..dependencies import require_fresh_token
-
-            token = await require_fresh_token(request, db)
-        except HTTPException:
-            token = None
-            logger.warning("No Graph API token available — creating contacts without sending")
-
-    sent = []
-    failed = []
-
-    if token and not is_testing:
-        # Real email send via Graph API
-        vendor_groups = []
-        for name, email in zip(vendor_names, vendor_emails):
-            if not email:
-                continue
-            vendor_groups.append(
-                {
-                    "vendor_name": name,
-                    "vendor_email": email,
-                    "parts": parts_text,
-                    "subject": subject,
-                    "body": body
-                    or f"Dear {name},\n\nWe are looking for the following parts: {parts_text}\n\nPlease provide your best pricing and availability.\n\nThank you.",
-                }
-            )
-
-        if vendor_groups:
-            try:
-                from ..email_service import send_batch_rfq
-
-                results = await send_batch_rfq(
-                    token=token,
-                    db=db,
-                    user_id=user.id,
-                    requisition_id=req_id,
-                    vendor_groups=vendor_groups,
-                )
-                for r in results:
-                    status = r.get("status", "sent")
-                    entry = {"vendor": r.get("vendor_name", ""), "email": r.get("vendor_email", ""), "status": status}
-                    if status == "failed":
-                        failed.append(entry)
-                    else:
-                        sent.append(entry)
-            except Exception as exc:
-                logger.error("Batch RFQ send failed: {}", exc)
-                # Fall back to DB-only mode
-                for name, email in zip(vendor_names, vendor_emails):
-                    if not email:
-                        continue
-                    contact = RfqContact(
-                        requisition_id=req_id,
-                        user_id=user.id,
-                        contact_type="email",
-                        vendor_name=name,
-                        vendor_name_normalized=name.lower().strip(),
-                        vendor_contact=email,
-                        parts_included=parts_text,
-                        subject=subject,
-                        status=ContactStatus.PENDING,
-                        status_updated_at=datetime.now(timezone.utc),
-                    )
-                    db.add(contact)
-                    sent.append({"vendor": name, "email": email, "status": "draft"})
-                db.commit()
-    else:
-        # Test mode or no token — create Contact records without sending
-        for name, email in zip(vendor_names, vendor_emails):
-            if not email:
-                continue
-            contact = RfqContact(
-                requisition_id=req_id,
-                user_id=user.id,
-                contact_type="email",
-                vendor_name=name,
-                vendor_name_normalized=name.lower().strip(),
-                vendor_contact=email,
-                parts_included=parts_text,
-                subject=subject,
-                status=ContactStatus.SENT,
-                status_updated_at=datetime.now(timezone.utc),
-            )
-            db.add(contact)
-            sent.append({"vendor": name, "email": email, "status": "sent"})
-        db.commit()
-
-    logger.info("RFQ: {} sent, {} failed for req {} by {}", len(sent), len(failed), req_id, user.email)
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req"] = req
-    ctx["sent_results"] = sent
-    ctx["failed_results"] = failed
-    ctx["total_sent"] = len(sent)
-    ctx["total_failed"] = len(failed)
-    return template_response("htmx/partials/requisitions/rfq_results.html", ctx)
-
-
-# ── Follow-ups & Response Review (Phase 6) ───────────────────────────
-
-
-@router.get("/v2/partials/follow-ups", response_class=HTMLResponse)
-async def follow_ups_list_partial(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Cross-requisition follow-up queue as HTML partial."""
-    from ..config import settings as cfg
-    from ..models.offers import Contact as RfqContact
-
-    threshold_days = getattr(cfg, "follow_up_days", 2)
-    threshold = datetime.now() - __import__("datetime").timedelta(days=threshold_days)
-
-    stale_q = db.query(RfqContact).filter(
-        RfqContact.contact_type == "email",
-        RfqContact.status.in_(["sent", "opened"]),
-        RfqContact.created_at < threshold,
-    )
-    if getattr(user, "role", None) in (UserRole.SALES, UserRole.TRADER):
-        stale_q = stale_q.join(Requisition).filter(Requisition.created_by == user.id)
-
-    stale = stale_q.order_by(RfqContact.created_at.asc()).limit(500).all()
-
-    req_ids = {c.requisition_id for c in stale}
-    req_names: dict[int, str] = {}
-    if req_ids:
-        for r in db.query(Requisition.id, Requisition.name).filter(Requisition.id.in_(req_ids)).all():
-            req_names[r.id] = r.name
-
-    from datetime import timezone as tz
-
-    now = datetime.now(tz.utc)
-    follow_ups = []
-    for c in stale:
-        ca = c.created_at if c.created_at else now
-        days_waiting = (now - ca).days
-        follow_ups.append(
-            {
-                "contact_id": c.id,
-                "requisition_id": c.requisition_id,
-                "requisition_name": req_names.get(c.requisition_id, "Unknown"),
-                "vendor_name": c.vendor_name,
-                "vendor_email": c.vendor_contact,
-                "parts": c.parts_included or [],
-                "status": c.status,
-                "days_waiting": days_waiting,
-            }
-        )
-
-    ctx = _base_ctx(request, user, "follow-ups")
-    ctx.update({"follow_ups": follow_ups, "total": len(follow_ups)})
-    return template_response("htmx/partials/follow_ups/list.html", ctx)
-
-
-@router.post("/v2/partials/follow-ups/{contact_id}/send", response_class=HTMLResponse)
-async def send_follow_up_htmx(
-    request: Request,
-    contact_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Send a follow-up email for a stale contact.
-
-    Returns success card.
-    """
-    import os
-
-    from ..models.offers import Contact as RfqContact
-
-    contact = db.get(RfqContact, contact_id)
-    if not contact:
-        raise HTTPException(404, "Contact not found")
-    require_requisition_access(db, contact.requisition_id, user, owner_id=contact.user_id, label="Contact")
-
-    form = await request.form()
-    body = (form.get("body") or "").strip()
-
-    # DNC hard-block — never email a do-not-contact vendor (checked in all modes,
-    # before the TESTING gate), mirroring send_reply_htmx / send_batch_rfq.
-    if contact.vendor_contact:
-        dnc = (
-            db.query(SiteContact)
-            .filter(
-                sqlfunc.lower(SiteContact.email) == contact.vendor_contact.lower(),
-                SiteContact.do_not_contact.is_(True),
-            )
-            .first()
-        )
-        if dnc:
-            logger.warning(
-                "Follow-up skipped — do-not-contact flag set for vendor '{}' ({})",
-                contact.vendor_name,
-                contact.vendor_contact,
-            )
-            return HTMLResponse(
-                '<div class="rounded bg-rose-50 border border-rose-200 text-rose-700 text-xs px-2 py-1.5">'
-                "This vendor is on the do-not-contact list — follow-up not sent.</div>"
-            )
-
-    is_testing = os.environ.get("TESTING") == "1"
-    email_sent = False
-
-    if not is_testing and contact.vendor_contact:
-        # Try to send real follow-up via Graph API
-        try:
-            from ..dependencies import require_fresh_token
-
-            token = await require_fresh_token(request, db)
-
-            from ..utils.graph_client import GraphClient
-
-            gc = GraphClient(token)
-            follow_up_subject = f"Follow-up: {contact.subject or 'RFQ'}"
-            follow_up_body = (
-                body
-                or f"Dear {contact.vendor_name},\n\nI'm following up on our previous inquiry. Please let us know if you have availability.\n\nThank you."
-            )
-            payload = {
-                "message": {
-                    "subject": follow_up_subject,
-                    "body": {"contentType": "Text", "content": follow_up_body},
-                    "toRecipients": [{"emailAddress": {"address": contact.vendor_contact}}],
-                },
-                "saveToSentItems": "true",
-            }
-            await gc.post_json("/me/sendMail", payload)
-            email_sent = True
-        except Exception as exc:
-            logger.warning("Follow-up email send failed for contact {}: {}", contact_id, exc)
-
-    from datetime import timezone as tz
-
-    if email_sent or is_testing:
-        contact.status = ContactStatus.SENT
-        contact.status_updated_at = datetime.now(tz.utc)
-        db.commit()
-
-    mode = "via Graph API" if email_sent else ("test mode" if is_testing else "FAILED")
-    logger.info(
-        "Follow-up {} for contact {} (vendor: {}, {}) by {}",
-        "sent" if email_sent or is_testing else "FAILED",
-        contact_id,
-        contact.vendor_name,
-        mode,
-        user.email,
-    )
-
-    ctx = _base_ctx(request, user, "follow-ups")
-    ctx["contact_id"] = contact_id
-    ctx["vendor_name"] = contact.vendor_name or "Vendor"
-    return template_response("htmx/partials/follow_ups/sent_success.html", ctx)
-
-
-@router.post("/v2/partials/follow-ups/{contact_id}/ai-draft", response_class=HTMLResponse)
-async def ai_draft_follow_up(
-    request: Request,
-    contact_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Draft a contextual follow-up body and fill the compose textarea."""
-    from datetime import timezone as tz
-
-    from ..models.offers import Contact as RfqContact
-
-    contact = db.get(RfqContact, contact_id)
-    if not contact:
-        raise HTTPException(404, "Contact not found")
-    require_requisition_access(db, contact.requisition_id, user, owner_id=contact.user_id, label="Contact")
-
-    days_waiting = (datetime.now(tz.utc) - contact.created_at).days if contact.created_at else None
-
-    from app.services.email_drafting import draft_email
-
-    result = await draft_email(
-        "follow_up",
-        {
-            "vendor_name": contact.vendor_name,
-            "parts": contact.parts_included or [],
-            "days_waiting": days_waiting,
-            "subject": contact.subject,
-        },
-    )
-    drafted = (result or {}).get("body") or ""
-
-    escaped = drafted.replace("\\", "\\\\").replace("`", "\\`").replace("$", "\\$").replace("</", "<\\/")
-    return HTMLResponse(
-        f'<script>document.getElementById("follow-up-body-{contact_id}").value = `{escaped}`;</script>'
-        '<p class="text-xs text-green-600 mt-1">Draft ready. Review and edit before sending.</p>'
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/responses/{response_id}/review", response_class=HTMLResponse)
-async def review_response_htmx(
-    request: Request,
-    req_id: int,
-    response_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Mark a vendor response as reviewed or rejected.
-
-    Returns updated card.
-    """
-    from ..models.offers import VendorResponse
-
-    require_requisition_access(db, req_id, user)
-    vr = (
-        db.query(VendorResponse)
-        .filter(
-            VendorResponse.id == response_id,
-            VendorResponse.requisition_id == req_id,
-        )
-        .first()
-    )
-    if not vr:
-        raise HTTPException(404, "Response not found")
-
-    form = await request.form()
-    new_status = form.get("status", "")
-    if new_status not in ("reviewed", "rejected"):
-        raise HTTPException(400, "Status must be 'reviewed' or 'rejected'")
-
-    vr.status = new_status
-    db.commit()
-    logger.info("Response {} marked as {} by {}", response_id, new_status, user.email)
-
-    req = db.query(Requisition).filter(Requisition.id == req_id).first()
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["r"] = vr
-    ctx["req"] = req
-    return template_response("htmx/partials/requisitions/tabs/response_card.html", ctx)
-
-
-@router.post(
-    "/v2/partials/requisitions/{req_id}/responses/{response_id}/ai-draft-reply",
-    response_class=HTMLResponse,
-)
-async def ai_draft_reply(
-    request: Request,
-    req_id: int,
-    response_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Draft an AI reply to a vendor response and render an editable compose block."""
-    from ..models.offers import VendorResponse
-
-    require_requisition_access(db, req_id, user)
-
-    vr = (
-        db.query(VendorResponse)
-        .filter(
-            VendorResponse.id == response_id,
-            VendorResponse.requisition_id == req_id,
-        )
-        .first()
-    )
-    if not vr:
-        raise HTTPException(404, "Response not found")
-
-    parsed = vr.parsed_data or {}
-
-    from app.services.email_drafting import draft_email
-
-    result = await draft_email(
-        "vendor_reply",
-        {
-            "classification": vr.classification,
-            "vendor_name": vr.vendor_name,
-            "mpn": parsed.get("mpn"),
-            "qty": parsed.get("qty") or parsed.get("qty_available"),
-            "price": parsed.get("price") or parsed.get("unit_price"),
-            "lead_time": parsed.get("lead_time"),
-            "subject": vr.subject,
-        },
-    )
-
-    default_subject = vr.subject or "RFQ"
-    if not default_subject.lower().startswith("re:"):
-        default_subject = f"Re: {default_subject}"
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["req_id"] = req_id
-    ctx["r"] = vr
-    ctx["reply_subject"] = (result or {}).get("subject") or default_subject
-    ctx["reply_body"] = (result or {}).get("body") or ""
-    ctx["ai_failed"] = result is None
-    return template_response("htmx/partials/requisitions/tabs/reply_compose.html", ctx)
-
-
-@router.post(
-    "/v2/partials/requisitions/{req_id}/responses/{response_id}/send-reply",
-    response_class=HTMLResponse,
-)
-async def send_reply_htmx(
-    request: Request,
-    req_id: int,
-    response_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Send a reply to a vendor response (as the signed-in user) and mark it
-    reviewed."""
-    import os
-
-    from ..models.offers import VendorResponse
-
-    require_requisition_access(db, req_id, user)
-
-    vr = (
-        db.query(VendorResponse)
-        .filter(
-            VendorResponse.id == response_id,
-            VendorResponse.requisition_id == req_id,
-        )
-        .first()
-    )
-    if not vr:
-        raise HTTPException(404, "Response not found")
-
-    form = await request.form()
-    subject = (form.get("subject") or "").strip() or f"Re: {vr.subject or 'RFQ'}"
-    body = (form.get("body") or "").strip()
-    if not body:
-        raise HTTPException(400, "Reply body is required")
-
-    # DNC hard-block — never email a do-not-contact vendor (checked in all modes).
-    if vr.vendor_email:
-        from sqlalchemy import func as _sqlfunc
-
-        dnc = (
-            db.query(SiteContact)
-            .filter(
-                _sqlfunc.lower(SiteContact.email) == vr.vendor_email.lower(),
-                SiteContact.do_not_contact.is_(True),
-            )
-            .first()
-        )
-        if dnc:
-            return HTMLResponse(
-                '<div class="rounded bg-rose-50 border border-rose-200 text-rose-700 text-xs px-2 py-1.5">'
-                "This vendor is on the do-not-contact list — reply not sent.</div>"
-            )
-
-    is_testing = os.environ.get("TESTING") == "1"
-    email_sent = False
-
-    if not is_testing and vr.vendor_email:
-        try:
-            from ..dependencies import require_fresh_token
-
-            token = await require_fresh_token(request, db)
-
-            from ..utils.graph_client import GraphClient
-
-            gc = GraphClient(token)
-            payload = {
-                "message": {
-                    "subject": subject,
-                    "body": {"contentType": "Text", "content": body},
-                    "toRecipients": [{"emailAddress": {"address": vr.vendor_email}}],
-                },
-                "saveToSentItems": "true",
-            }
-            await gc.post_json("/me/sendMail", payload)
-            email_sent = True
-        except Exception as exc:
-            logger.warning("Vendor reply send failed for response {}: {}", response_id, exc)
-
-    if email_sent or is_testing:
-        vr.status = VendorResponseStatus.REVIEWED
-        db.commit()
-
-    logger.info(
-        "Vendor reply {} for response {} (vendor: {}) by {}",
-        "sent" if email_sent or is_testing else "FAILED",
-        response_id,
-        vr.vendor_name,
-        user.email,
-    )
-
-    req = db.query(Requisition).filter(Requisition.id == req_id).first()
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["r"] = vr
-    ctx["req"] = req
-    return template_response("htmx/partials/requisitions/tabs/response_card.html", ctx)
 
 
 @router.post("/v2/partials/requisitions/{req_id}/poll-inbox", response_class=HTMLResponse)
@@ -3269,238 +1669,6 @@ async def edit_quote_metadata(
     return await quote_detail_partial(request=request, quote_id=quote_id, user=user, db=db)
 
 
-# ── Sprint 6: RFQ Workflow Depth ────────────────────────────────────────
-
-
-@router.get("/v2/partials/requisitions/{req_id}/rfq-prepare", response_class=HTMLResponse)
-async def rfq_prepare_panel(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return RFQ preparation panel — vendor data + exhaustion check."""
-    from ..models.offers import Contact as RfqContact
-
-    req = get_requisition_or_404(db, req_id)
-
-    # Get requirements for this req
-    requirements = db.query(Requirement).filter(Requirement.requisition_id == req_id).all()
-    mpns = [r.primary_mpn for r in requirements if r.primary_mpn]
-
-    # Get vendors already contacted
-    existing_contacts = (
-        db.query(RfqContact.vendor_name_normalized).filter(RfqContact.requisition_id == req_id).distinct().all()
-    )
-    contacted_norms = {c[0] for c in existing_contacts if c[0]}
-
-    # Get suggested vendors from sightings (join on normalized vendor name)
-    from ..models import Sighting
-
-    suggested_vendors = (
-        (
-            db.query(
-                VendorCard.id,
-                VendorCard.display_name,
-                VendorCard.normalized_name,
-                sqlfunc.count(Sighting.id).label("sighting_count"),
-            )
-            .join(Sighting, Sighting.vendor_name_normalized == VendorCard.normalized_name)
-            .filter(
-                Sighting.mpn_matched.in_(mpns) if mpns else sqlfunc.literal(False),
-                VendorCard.is_blacklisted.isnot(True),
-            )
-            .group_by(VendorCard.id)
-            .order_by(sqlfunc.count(Sighting.id).desc())
-            .limit(20)
-            .all()
-        )
-        if mpns
-        else []
-    )
-
-    vendors = []
-    for v in suggested_vendors:
-        vendors.append(
-            {
-                "id": v.id,
-                "display_name": v.display_name,
-                "normalized_name": v.normalized_name,
-                "sighting_count": v.sighting_count,
-                "already_contacted": v.normalized_name in contacted_norms,
-            }
-        )
-
-    return template_response(
-        "htmx/partials/requisitions/rfq_prepare.html",
-        {"request": request, "req": req, "vendors": vendors, "mpns": mpns, "total_contacted": len(contacted_norms)},
-    )
-
-
-@router.post("/v2/partials/requisitions/{req_id}/log-phone", response_class=HTMLResponse)
-async def log_phone_call(
-    request: Request,
-    req_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Log a phone call to a vendor and return updated activity tab."""
-    get_requisition_or_404(db, req_id)  # validates existence
-    require_requisition_access(db, req_id, user)
-
-    form = await request.form()
-    vendor_name = form.get("vendor_name", "").strip()
-    vendor_phone = form.get("vendor_phone", "").strip()
-    notes = form.get("notes", "").strip()
-
-    if not vendor_name or not vendor_phone:
-        raise HTTPException(400, "Vendor name and phone are required")
-
-    from ..models.offers import Contact as RfqContact
-    from ..services.activity_service import log_call_activity
-
-    contact = RfqContact(
-        requisition_id=req_id,
-        user_id=user.id,
-        contact_type="phone",
-        vendor_name=vendor_name,
-        vendor_contact=vendor_phone,
-        details=notes or f"Phone call to {vendor_name}",
-        status=ContactStatus.SENT,
-    )
-    db.add(contact)
-
-    # Route through log_call_activity so the call is matched to a vendor/company,
-    # recorded as the canonical CALL_LOGGED type, and bumps last_activity_at.
-    # force_meaningful=True: a manually logged call is a deliberate human interaction →
-    # always meaningful, regardless of the duration-gate (which targets auto-captured
-    # Teams/8x8 calls that carry a real duration).
-    log = log_call_activity(
-        user_id=user.id,
-        direction="outbound",
-        phone=vendor_phone,
-        duration_seconds=None,
-        external_id=None,
-        contact_name=vendor_name,
-        db=db,
-        requisition_id=req_id,
-        force_meaningful=True,
-    )
-    if log is not None:
-        log.notes = notes or f"Called {vendor_name} at {vendor_phone}"
-    db.commit()
-    logger.info("Phone call logged for req {} vendor {} by {}", req_id, vendor_name, user.email)
-
-    return template_response(
-        "htmx/partials/requisitions/phone_log_success.html",
-        {"request": request, "vendor_name": vendor_name, "vendor_phone": vendor_phone},
-    )
-
-
-@router.post("/v2/partials/follow-ups/send-batch", response_class=HTMLResponse)
-async def send_batch_follow_up(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Send follow-ups to all stale contacts at once."""
-    from ..models.offers import Contact as RfqContact
-
-    cfg = getattr(request.app, "state", None)
-    threshold_days = getattr(cfg, "follow_up_days", 2) if cfg else 2
-    threshold = datetime.now(timezone.utc) - timedelta(days=threshold_days)
-
-    q = db.query(RfqContact).filter(
-        RfqContact.contact_type == "email",
-        RfqContact.status.in_(["sent", "opened"]),
-        RfqContact.created_at < threshold,
-    )
-    # Restricted roles act only on contacts under their own requisitions; buyer/manager/admin
-    # stay global. Keep this in lockstep with follow_up_badge so the badge counts what the
-    # batch acts on.
-    if user.role in RESTRICTED_ROLES:
-        q = q.join(Requisition, RfqContact.requisition_id == Requisition.id).filter(Requisition.created_by == user.id)
-    stale = q.limit(50).all()
-
-    sent_count = 0
-    for contact in stale:
-        contact.status = ContactStatus.RESPONDED
-        contact.status_updated_at = datetime.now(timezone.utc)
-        sent_count += 1
-    db.commit()
-    logger.info("Batch follow-up: {} contacts marked by {}", sent_count, user.email)
-
-    msg = f"{sent_count} contact{'s' if sent_count != 1 else ''} marked as responded."
-    return HTMLResponse(
-        f'<div class="text-sm text-green-700 bg-green-50 border border-green-200 rounded-lg px-3 py-2">{msg}</div>'
-    )
-
-
-@router.get("/v2/partials/follow-ups/badge", response_class=HTMLResponse)
-async def follow_up_badge(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return follow-up count badge for nav sidebar."""
-    from ..models.offers import Contact as RfqContact
-
-    threshold = datetime.now(timezone.utc) - timedelta(days=2)
-    q = db.query(sqlfunc.count(RfqContact.id)).filter(
-        RfqContact.contact_type == "email",
-        RfqContact.status.in_(["sent", "opened"]),
-        RfqContact.created_at < threshold,
-    )
-    # Same per-owner scope as send_batch_follow_up so the badge matches the batch.
-    if user.role in RESTRICTED_ROLES:
-        q = q.join(Requisition, RfqContact.requisition_id == Requisition.id).filter(Requisition.created_by == user.id)
-    count = q.scalar() or 0
-    if count > 0:
-        return HTMLResponse(
-            f'<span class="ml-auto px-1.5 py-0.5 text-[10px] font-bold text-white bg-amber-500 rounded-full">{count}</span>'
-        )
-    return HTMLResponse("")
-
-
-@router.patch("/v2/partials/requisitions/{req_id}/responses/{response_id}/status", response_class=HTMLResponse)
-async def update_response_status(
-    request: Request,
-    req_id: int,
-    response_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Update vendor response status (reviewed/rejected/flagged)."""
-    from ..models.offers import VendorResponse
-
-    require_requisition_access(db, req_id, user)
-    vr = (
-        db.query(VendorResponse)
-        .filter(
-            VendorResponse.id == response_id,
-            VendorResponse.requisition_id == req_id,
-        )
-        .first()
-    )
-    if not vr:
-        raise HTTPException(404, "Response not found")
-
-    form = await request.form()
-    new_status = form.get("status", "").strip()
-    valid = {"reviewed", "rejected", "flagged", "new"}
-    if new_status not in valid:
-        raise HTTPException(400, f"Invalid status. Must be one of: {', '.join(valid)}")
-
-    vr.status = new_status
-    db.commit()
-    logger.info("Response {} status → {} by {}", response_id, new_status, user.email)
-
-    return template_response(
-        "htmx/partials/requisitions/response_status_badge.html",
-        {"request": request, "response": vr},
-    )
-
-
 # ── Sprint 7: Email Integration ────────────────────────────────────────
 
 
@@ -3819,940 +1987,6 @@ async def pipeline_insights_refresh(
     return _render_insights(request, user, insights, "dashboard", 0)
 
 
-# ── Buy Plans partials ─────────────────────────────────────────────────
-
-
-def _is_ops_member(user: User, db: Session) -> bool:
-    """Check if user is in the ops verification group."""
-    return db.query(VerificationGroupMember).filter_by(user_id=user.id, is_active=True).first() is not None
-
-
-def _can_supervise(user: User, db: Session) -> bool:
-    """True when the user may see cross-user (scope=all) deal data.
-
-    Managers/admins and ops verification-group members qualify.
-    """
-    return user.role in (UserRole.MANAGER, UserRole.ADMIN) or _is_ops_member(user, db)
-
-
-# Roles that cut/claim POs. Deliberately NOT the broader BUYER_ROLES (which includes
-# SALES/TRADER) — only these may re-source a cancelled PO or claim an open-pool line.
-_PO_CUTTER_ROLES = (UserRole.BUYER, UserRole.MANAGER, UserRole.ADMIN)
-
-
-def _can_resource(user: User) -> bool:
-    """True when the user may re-source / claim buy-plan lines (a PO-cutter)."""
-    return user.role in _PO_CUTTER_ROLES
-
-
-def _can_see_all_deals(user: User, db: Session) -> bool:
-    """True when the user may view every owner's deals on the Deal Hub board.
-
-    PO-cutters (buyers + managers/admins) and ops verification-group members see the
-    full deal flow; sales/traders are scoped to their own deals only. Broader than
-    ``_can_supervise`` by including buyers, who need cross-owner visibility to cut POs.
-    """
-    return _can_resource(user) or _is_ops_member(user, db)
-
-
-def _resolve_deal_scope(scope: str, can_see_all: bool) -> str:
-    """Normalize a requested deal scope against the user's visibility.
-
-    Empty/unknown → the role default (``all`` for can-see-all users, else ``mine``).
-    ``all`` requested by a user without cross-owner visibility is forced to ``mine`` so
-    no other rep's plans leak.
-    """
-    if scope not in ("mine", "all"):
-        return "all" if can_see_all else "mine"
-    if scope == "all" and not can_see_all:
-        return "mine"
-    return scope
-
-
-def _require_po_cutter(user: User) -> None:
-    """403 unless the user is an active PO-cutter (buyer/manager/admin)."""
-    if not _can_resource(user) or not getattr(user, "is_active", True):
-        raise HTTPException(403, "Only buyers and managers can re-source / claim lines")
-
-
-# Canonical stage-tab lens keys (underscored). URL paths use dashes (sales-orders, …).
-_APPROVALS_TABS = ("sales_orders", "buy_plans", "purchase_orders", "prepayments", "supervise")
-
-# Per-gate approve-right attribute that gates each stage tab's pinned "Pending approvals"
-# section. The keys match services.approvals.queue.TAB_GATE; the values are User columns.
-# buy_plans is intentionally absent — that tab is gate-less (board only, no pending section).
-_TAB_APPROVE_ATTR = {
-    "sales_orders": "can_approve_buy_plans",
-    "purchase_orders": "can_approve_pos",
-    "prepayments": "can_approve_prepayments",
-}
-
-
-def _default_lens(user: User, db: Session) -> str:
-    """Pick the landing stage tab for the Approvals hub based on the user's role.
-
-    - buyers land on the Purchase Orders stage (their PO cut queue),
-    - managers/admins/ops land on Supervise,
-    - everyone else (sales/trader) lands on the Buy Plans deal board.
-    """
-    if user.role == UserRole.BUYER:
-        return "purchase_orders"
-    if _can_supervise(user, db):
-        return "supervise"
-    return "buy_plans"
-
-
-@router.get("/v2/partials/approvals", response_class=HTMLResponse)
-@router.get("/v2/partials/buy-plans", response_class=HTMLResponse)
-async def buy_plans_list_partial(
-    request: Request,
-    lens: str = "",
-    user: User = Depends(require_access(AccessKey.BUY_PLANS)),
-    db: Session = Depends(get_db),
-):
-    """Return the Approvals hub shell (stage-tab switcher).
-
-    The shell renders the five lifecycle stage tabs + a lazy body that loads the active
-    stage tab partial into ``#bp-hub-body``. Row data is fetched by the body, not here.
-    ``/v2/partials/buy-plans`` is kept as a back-compat alias for in-flight htmx.
-    """
-    active_lens = lens if lens in _APPROVALS_TABS else _default_lens(user, db)
-
-    # Spotlight markers: plan rows that carry an open step needing this user's action.
-    # Buy Plans is its own primary nav tab, so the source is registered under "buy-plans".
-    from ..services.alerts import markers_for_tab
-
-    alert_markers = markers_for_tab(db, user, "buy-plans")
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx.update(
-        {
-            "lens": active_lens,
-            "alert_markers": alert_markers,
-            # Only Supervise is gate-rendered in the shell; the four stage tabs are always
-            # shown (their work surface + pinned approval section gate by role inside).
-            "can_supervise": _can_supervise(user, db),
-        }
-    )
-    return template_response("htmx/partials/buy_plans/hub.html", ctx)
-
-
-@router.get("/v2/partials/approvals/{tab}", response_class=HTMLResponse)
-async def approvals_tab_partial(
-    request: Request,
-    tab: str,
-    scope: str = "",
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Render one Approvals stage-tab body into ``#bp-hub-body``.
-
-    Composes the re-homed work surface for the stage (deal board / buyer orders +
-    re-sourcing pool / neutral empty state) with a pinned per-gate "Pending approvals"
-    section (services.approvals.queue.build_queue_view), shown only when the viewer holds
-    that gate's approve right. ``supervise`` reuses the manager triage body. ``tab`` arrives
-    dash-cased (e.g. purchase-orders) and maps to the underscored stage key.
-
-    ``scope`` applies to the Buy Plans stage's deal board only: it is role-resolved exactly
-    like the standalone board (sales/traders locked to ``mine``), and its All/Mine toggle
-    reloads THIS whole tab body so the pinned approval section survives the swap.
-    """
-    lens = tab.replace("-", "_")
-    if lens not in _APPROVALS_TABS:
-        raise HTTPException(404, "Unknown approvals tab")
-
-    if lens == "supervise":
-        return _render_supervise_body(request, user, db)
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    if lens in _TAB_APPROVE_ATTR:
-        from ..services.approvals.queue import build_queue_view
-
-        ctx["view"] = build_queue_view(db, user, lens)
-        ctx["show_pending"] = bool(getattr(user, _TAB_APPROVE_ATTR[lens], False))
-
-    if lens == "buy_plans":
-        from ..services.buyplan_hub import completed_archive, deals_board
-
-        # Role-resolve the deal-board scope exactly like the standalone /board route, but
-        # point the All/Mine toggle at THIS tab URL so a toggle reloads the whole tab body
-        # (pinned approval section + board) rather than swapping in the bare board.
-        can_all = _can_see_all_deals(user, db)
-        board_scope = _resolve_deal_scope(scope, can_all)
-        ctx.update(
-            {
-                "board": deals_board(
-                    db,
-                    user,
-                    scope=board_scope,
-                    statuses=[BuyPlanStatus.ACTIVE.value, BuyPlanStatus.HALTED.value],
-                ),
-                "scope": board_scope,
-                "archive": completed_archive(db, user, scope=board_scope),
-                "can_see_all_deals": can_all,
-                "scope_toggle_url": "/v2/partials/approvals/buy-plans",
-            }
-        )
-        return template_response("htmx/partials/approvals/_tab_buy_plans.html", ctx)
-
-    if lens == "purchase_orders":
-        from ..services.buyplan_hub import buyer_line_queue, resourcing_pool_queue, team_line_queue
-
-        ctx.update(
-            {
-                "orders_queue": buyer_line_queue(db, user),
-                "team": team_line_queue(db, user),
-                "resource_queue": resourcing_pool_queue(db),
-                "can_claim": _can_resource(user),
-            }
-        )
-        return template_response("htmx/partials/approvals/_tab_purchase_orders.html", ctx)
-
-    if lens == "sales_orders":
-        from ..services.buyplan_hub import deals_board
-
-        can_all = _can_see_all_deals(user, db)
-        board_scope = _resolve_deal_scope(scope, can_all)
-        # DRAFT/PENDING work surface only. No `archive`: the Completed archive belongs to
-        # the Buy Plans tab, and the shared board renders it only when `archive` is passed.
-        ctx.update(
-            {
-                "board": deals_board(
-                    db,
-                    user,
-                    scope=board_scope,
-                    statuses=[BuyPlanStatus.DRAFT.value, BuyPlanStatus.PENDING.value],
-                ),
-                "scope": board_scope,
-                "can_see_all_deals": can_all,
-                "scope_toggle_url": "/v2/partials/approvals/sales-orders",
-            }
-        )
-        return template_response("htmx/partials/approvals/_tab_sales_orders.html", ctx)
-
-    # prepayments — approval-only stage (no work surface in SP-1)
-    return template_response("htmx/partials/approvals/_tab_prepayments.html", ctx)
-
-
-@router.get("/v2/partials/approvals/sales-orders/new", response_class=HTMLResponse)
-async def sales_order_new(
-    request: Request,
-    requisition_id: int | None = None,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """New Sales Order origination surface (requisition picker → offer/sell builder).
-
-    The two-segment path after ``approvals/`` does not collide with the one-segment
-    ``/v2/partials/approvals/{tab}`` converter. With no ``requisition_id`` it lists open
-    (OPEN_PIPELINE) requisitions that carry at least one ACTIVE offer, scoped to what the
-    user may see. With ``requisition_id`` it loads that requisition's per-requirement
-    offer/sell-price form (``get_builder_data`` + ``apply_smart_defaults``), enforcing access
-    via ``get_req_for_user`` (404 for a restricted role that does not own it).
-    """
-    from sqlalchemy import func
-
-    from ..constants import OfferStatus, RequisitionStatus
-    from ..dependencies import get_req_for_user
-    from ..models import Offer, Requirement
-    from ..services.quote_builder_service import apply_smart_defaults, get_builder_data
-
-    ctx = _base_ctx(request, user, "buy-plans")
-
-    if requisition_id is not None:
-        req = get_req_for_user(db, user, requisition_id)
-        lines = get_builder_data(req.id, db)
-        apply_smart_defaults(lines)
-        ctx.update({"selected_req": req, "lines": lines})
-        return template_response("htmx/partials/approvals/_sales_order_new.html", ctx)
-
-    # Picker mode: open requisitions with at least one active offer, scoped to the viewer.
-    has_active_offer = (
-        select(Offer.id)
-        .join(Requirement, Offer.requirement_id == Requirement.id)
-        .where(
-            Requirement.requisition_id == Requisition.id,
-            Offer.status == OfferStatus.ACTIVE,
-        )
-        .exists()
-    )
-    q = db.query(Requisition).filter(
-        Requisition.status.in_(list(RequisitionStatus.OPEN_PIPELINE)),
-        has_active_offer,
-    )
-    if user.role in RESTRICTED_ROLES:
-        q = q.filter(Requisition.created_by == user.id)
-    reqs = q.order_by(Requisition.id.desc()).all()
-
-    counts: dict[int, int] = {}
-    if reqs:
-        counts = dict(
-            db.query(Requirement.requisition_id, func.count(Offer.id))
-            .join(Offer, Offer.requirement_id == Requirement.id)
-            .filter(
-                Requirement.requisition_id.in_([r.id for r in reqs]),
-                Offer.status == OfferStatus.ACTIVE,
-            )
-            .group_by(Requirement.requisition_id)
-            .all()
-        )
-
-    picker_rows = [
-        {"id": r.id, "name": r.name, "customer": r.customer_name or "", "offer_count": counts.get(r.id, 0)}
-        for r in reqs
-    ]
-    ctx.update({"selected_req": None, "picker_rows": picker_rows})
-    return template_response("htmx/partials/approvals/_sales_order_new.html", ctx)
-
-
-@router.post("/v2/partials/approvals/sales-orders/create", response_class=HTMLResponse)
-async def sales_order_create(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Originate a DRAFT buy plan (Sales Order) from the chosen offers, then render its
-    detail.
-
-    Parses ``requisition_id`` + per-requirement ``offer_<rid>`` / ``sell_<rid>`` form fields,
-    enforces requisition access (``require_requisition_access`` — 404 for a restricted role
-    that does not own it), and calls ``create_sales_order_from_offers``. On the builder's
-    duplicate-open-SO ValueError it renders the existing open Sales Order's detail with a
-    toast (never a 500); any other ValueError (e.g. no requirements) is a 400.
-    """
-    from ..dependencies import require_requisition_access
-    from ..services.buyplan_builder import (
-        DuplicateSalesOrderError,
-        create_sales_order_from_offers,
-    )
-
-    form = await request.form()
-    raw_req_id = form.get("requisition_id")
-    if not raw_req_id:
-        raise HTTPException(400, "Requisition is required")
-    try:
-        req_id = int(raw_req_id)
-    except (TypeError, ValueError):
-        raise HTTPException(400, "Invalid requisition")
-
-    require_requisition_access(db, req_id, user)
-
-    selections: dict[int, int] = {}
-    sell_prices: dict[int, float] = {}
-    for key, value in form.multi_items():
-        if key.startswith("offer_"):
-            try:
-                selections[int(key[len("offer_") :])] = int(value)
-            except (TypeError, ValueError):
-                continue
-        elif key.startswith("sell_"):
-            if value in (None, ""):
-                continue
-            try:
-                sell_prices[int(key[len("sell_") :])] = float(value)
-            except (TypeError, ValueError):
-                continue
-
-    try:
-        plan = create_sales_order_from_offers(req_id, selections, sell_prices, db, user)
-    except DuplicateSalesOrderError as exc:
-        # An open Sales Order already exists for this requisition — open it instead of
-        # 500ing. The exception carries the existing plan id, so no re-query is needed.
-        existing_id = exc.existing_plan_id
-        resp = await buy_plan_detail_partial(request, existing_id, user, db)
-        resp.headers["HX-Trigger"] = json.dumps(
-            {
-                "showToast": {
-                    "message": f"There is already an open buy plan for this requisition (plan #{existing_id}).",
-                    "type": "warning",
-                }
-            }
-        )
-        resp.headers["HX-Push-Url"] = f"/v2/buy-plans/{existing_id}"
-        return resp
-    except ValueError:
-        # Any other origination failure (e.g. requisition has no requirements). Return a
-        # curated client message rather than echoing the raw builder error.
-        raise HTTPException(400, "Could not build a buy plan from the selected offers.")
-
-    resp = await buy_plan_detail_partial(request, plan.id, user, db)
-    resp.headers["HX-Push-Url"] = f"/v2/buy-plans/{plan.id}"
-    return resp
-
-
-@router.get("/v2/partials/buy-plans/resource", response_class=HTMLResponse)
-async def buy_plans_resource_partial(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Open-claim queue body for the "Needs Re-sourcing" lens (pool-wide).
-
-    Lists every line whose cut PO was cancelled (vendor fell down) and is unassigned,
-    awaiting any PO-cutter to claim + backfill.
-    """
-    from ..services.buyplan_hub import resourcing_pool_queue
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx["queue"] = resourcing_pool_queue(db)
-    ctx["can_claim"] = _can_resource(user)
-    return template_response("htmx/partials/buy_plans/_resource_queue.html", ctx)
-
-
-@router.get("/v2/partials/buy-plans/orders", response_class=HTMLResponse)
-async def buy_plans_orders_partial(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Buyer Orders body (re-homed under the Purchase Orders stage tab): the actionable
-    per-line PO cut queue.
-
-    Also includes a read-only "Team Orders" awareness section listing open lines
-    assigned to OTHER buyers (see ``team_line_queue``).
-    """
-    from ..services.buyplan_hub import buyer_line_queue, team_line_queue
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx["queue"] = buyer_line_queue(db, user)
-    ctx["team"] = team_line_queue(db, user)
-    return template_response("htmx/partials/buy_plans/_orders_queue.html", ctx)
-
-
-@router.get("/v2/partials/buy-plans/board", response_class=HTMLResponse)
-async def buy_plans_board_partial(
-    request: Request,
-    scope: str = "",
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Deal board body (re-homed under the Buy Plans / Supervise stage tabs): stage-
-    grouped deal cards.
-
-    Scope is role-defaulted: PO-cutters + ops (``_can_see_all_deals``) default to
-    ``all`` and may toggle to ``mine``; sales/traders are locked to ``mine`` so no
-    other rep's plans leak.
-    """
-    from ..services.buyplan_hub import completed_archive, deals_board
-
-    can_all = _can_see_all_deals(user, db)
-    scope = _resolve_deal_scope(scope, can_all)
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx.update(
-        {
-            "board": deals_board(db, user, scope=scope),
-            "scope": scope,
-            "archive": completed_archive(db, user, scope=scope),
-            "can_see_all_deals": can_all,
-        }
-    )
-    return template_response("htmx/partials/buy_plans/_board.html", ctx)
-
-
-@router.get("/v2/partials/buy-plans/archive", response_class=HTMLResponse)
-async def buy_plans_archive_partial(
-    request: Request,
-    scope: str = "",
-    offset: int = 0,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Completed-transactions archive page (lazy "load older" chunk).
-
-    Returns just the rows partial (not the whole section) so an htmx "Load older" click
-    can append the next page in place. Scope is role-resolved exactly like the board so
-    no other rep's completed plans leak to a sales/trader user.
-    """
-    from ..services.buyplan_hub import completed_archive
-
-    scope = _resolve_deal_scope(scope, _can_see_all_deals(user, db))
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx.update(
-        {
-            "archive": completed_archive(db, user, scope=scope, offset=offset),
-            "scope": scope,
-        }
-    )
-    return template_response("htmx/partials/buy_plans/_archive_rows.html", ctx)
-
-
-def _render_supervise_body(request: Request, user: User, db: Session) -> HTMLResponse:
-    """Build + render the supervise lens body for ``user``.
-
-    Shared by the ``GET /supervise`` route and the supervise-origin action returns.
-    Non-supervisors never see cross-user data: they get the mine-scope board instead
-    (defense in depth — the hub also hides the Supervise button for them).
-    """
-    from ..services.buyplan_hub import completed_archive, deals_board, supervise_overview
-
-    if not _can_supervise(user, db):
-        ctx = _base_ctx(request, user, "buy-plans")
-        ctx.update(
-            {
-                "board": deals_board(db, user, scope="mine"),
-                "scope": "mine",
-                "archive": completed_archive(db, user, scope="mine"),
-            }
-        )
-        return template_response("htmx/partials/buy_plans/_board.html", ctx)
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx.update(
-        {
-            "overview": supervise_overview(db),
-            "board": deals_board(db, user, scope="all"),
-            "archive": completed_archive(db, user, scope="all"),
-            "is_ops": _is_ops_member(user, db),
-            "is_manager": user.role in (UserRole.MANAGER, UserRole.ADMIN),
-            "can_approve": can_approve_buy_plans(user),
-            "user": user,
-        }
-    )
-    return template_response("htmx/partials/buy_plans/_supervise.html", ctx)
-
-
-@router.get("/v2/partials/buy-plans/supervise", response_class=HTMLResponse)
-async def buy_plans_supervise_partial(
-    request: Request,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Manager/ops "Supervise" lens body: triage panel + all-scope deal board.
-
-    Role-gated — a non-supervisor is served the mine-scope board so no other
-    user's plans leak (see ``_render_supervise_body``).
-    """
-    return _render_supervise_body(request, user, db)
-
-
-@router.get("/v2/partials/buy-plans/{plan_id}", response_class=HTMLResponse)
-async def buy_plan_detail_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return buy plan detail as HTML partial."""
-    bp = get_buyplan_for_user(
-        db,
-        user,
-        plan_id,
-        options=[
-            joinedload(BuyPlan.lines).joinedload(BuyPlanLine.offer),
-            joinedload(BuyPlan.lines).joinedload(BuyPlanLine.requirement),
-            joinedload(BuyPlan.lines).joinedload(BuyPlanLine.buyer),
-            joinedload(BuyPlan.quote),
-            joinedload(BuyPlan.requisition),
-            joinedload(BuyPlan.submitted_by),
-            joinedload(BuyPlan.approved_by),
-        ],
-    )
-
-    ctx = _base_ctx(request, user, "buy-plans")
-    ctx.update(
-        {
-            "bp": bp,
-            "lines": bp.lines or [],
-            "is_ops_member": _is_ops_member(user, db),
-            "can_resource": _can_resource(user),
-            "user": user,
-            # Most-urgent flag reason so the indicator states the issue at first glance.
-            "top_flag": summarize_top_flag(bp.ai_flags),
-        }
-    )
-    return template_response("htmx/partials/buy_plans/detail.html", ctx)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/submit", response_class=HTMLResponse)
-async def buy_plan_submit_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Submit a draft buy plan with SO# — returns refreshed detail partial."""
-    from ..services.buyplan_notifications import (
-        notify_approved,
-        notify_submitted,
-        run_notify_bg,
-    )
-    from ..services.buyplan_workflow import submit_buy_plan
-
-    # Per-record ownership: non-owner SALES/TRADER → 404 before any mutation.
-    get_buyplan_for_user(db, user, plan_id)
-
-    form = await request.form()
-    so = form.get("sales_order_number", "").strip()
-    if not so:
-        raise HTTPException(400, "Sales Order # is required")
-
-    try:
-        plan = submit_buy_plan(
-            plan_id,
-            so,
-            user,
-            db,
-            customer_po_number=form.get("customer_po_number") or None,
-            salesperson_notes=form.get("salesperson_notes") or None,
-        )
-        db.commit()
-        if plan.auto_approved:
-            await run_notify_bg(notify_approved, plan.id)
-        else:
-            await run_notify_bg(notify_submitted, plan.id)
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/approve", response_class=HTMLResponse)
-async def buy_plan_approve_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_buyplan_approver),
-    db: Session = Depends(get_db),
-):
-    """Approve or reject a pending buy plan — returns refreshed detail.
-
-    Gated by ``require_buyplan_approver`` (403 unless the user holds the per-user
-    can_approve_buy_plans right). Reject requires a reason (enforced in the service).
-
-    QP Phase C1: the approval engine OWNS the gate. We look up the open BUY_PLAN
-    ApprovalRequest for this plan and resolve it via the engine's ``decide`` — which drives
-    the buy-plan side effects (ACTIVE + buyer tasks / DRAFT) in the SAME transaction. We let
-    ``decide`` raise (no swallowing) so a side-effect failure rolls back the whole decision
-    atomically (RISK 1). If NO open request exists — a plan that went PENDING before C1
-    deployed — we fall back to the legacy ``approve_buy_plan`` and log a WARNING (RISK 3,
-    transition window; the fallback is removed in a follow-up once no pre-C1 plans remain).
-    """
-    from sqlalchemy import select as _select
-
-    from ..constants import ApprovalRequestStatus, ApprovalSubjectType
-    from ..models.approvals import ApprovalRequest
-    from ..services.approvals.service import decide as svc_decide
-    from ..services.buyplan_notifications import (
-        notify_approved,
-        notify_rejected,
-        run_notify_bg,
-    )
-    from ..services.buyplan_workflow import approve_buy_plan
-
-    form = await request.form()
-    action = form.get("action", "approve")
-    origin = form.get("origin", "")
-    notes = form.get("notes")
-
-    open_request = (
-        db.execute(
-            _select(ApprovalRequest).where(
-                ApprovalRequest.subject_type == ApprovalSubjectType.BUY_PLAN,
-                ApprovalRequest.subject_id == plan_id,
-                ApprovalRequest.status == ApprovalRequestStatus.REQUESTED,
-            )
-        )
-        .scalars()
-        .first()
-    )
-
-    try:
-        if open_request is not None:
-            # Engine path: decide() resolves the request AND drives the plan side effects.
-            svc_decide(db, open_request.id, user, action, comment=notes or None)
-        else:
-            # RISK 3 fallback: plan pending pre-C1 with no engine request yet.
-            logger.warning(
-                "Buy plan {} approve/reject with no open engine request — falling back to legacy approve_buy_plan",
-                plan_id,
-            )
-            approve_buy_plan(plan_id, action, user, db, notes=notes)
-        db.commit()
-        if action == "approve":
-            await run_notify_bg(notify_approved, plan_id)
-        else:
-            await run_notify_bg(notify_rejected, plan_id)
-    except PermissionError as e:
-        # The dependency already 403s unauthorized callers; this maps the service's
-        # defense-in-depth approval-right check to 403 (not 400) if it is ever reached.
-        raise HTTPException(403, str(e))
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    if origin == "supervise":
-        return _render_supervise_body(request, user, db)
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/verify-so", response_class=HTMLResponse)
-async def buy_plan_verify_so_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Ops verifies SO — returns refreshed detail."""
-    from ..services.buyplan_notifications import (
-        notify_so_rejected,
-        notify_so_verified,
-        run_notify_bg,
-    )
-    from ..services.buyplan_workflow import verify_so
-
-    form = await request.form()
-    action = form.get("action", "approve")
-    origin = form.get("origin", "")
-
-    try:
-        plan = verify_so(
-            plan_id,
-            action,
-            user,
-            db,
-            rejection_note=form.get("rejection_note"),
-        )
-        db.commit()
-        if action == "approve":
-            await run_notify_bg(notify_so_verified, plan.id)
-        else:
-            await run_notify_bg(notify_so_rejected, plan.id, action=action)
-    except (ValueError, PermissionError) as e:
-        raise HTTPException(400, str(e))
-
-    if origin == "supervise":
-        return _render_supervise_body(request, user, db)
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/lines/{line_id}/confirm-po", response_class=HTMLResponse)
-async def buy_plan_confirm_po_partial(
-    request: Request,
-    plan_id: int,
-    line_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Buyer confirms PO.
-
-    Returns the refreshed detail partial by default (``origin=""``, the original
-    behavior). When ``origin == "queue"`` the call came from the buyer's Orders lens
-    and we return the re-rendered orders queue so the confirmed line drops out.
-    """
-    from datetime import datetime
-
-    from ..services.buyplan_notifications import notify_po_confirmed, run_notify_bg
-    from ..services.buyplan_workflow import confirm_po
-
-    # Per-record ownership: non-owner SALES/TRADER → 404 before any mutation.
-    get_buyplan_for_user(db, user, plan_id)
-
-    form = await request.form()
-    po_number = form.get("po_number", "").strip()
-    ship_date_str = form.get("estimated_ship_date", "")
-    origin = form.get("origin", "")
-
-    if not po_number:
-        raise HTTPException(400, "PO number is required")
-
-    ship_date = None
-    if ship_date_str:
-        try:
-            ship_date = datetime.fromisoformat(ship_date_str)
-        except ValueError:
-            ship_date = datetime.now()
-    else:
-        ship_date = datetime.now()
-
-    try:
-        confirm_po(plan_id, line_id, po_number, ship_date, user, db)
-        db.commit()
-        await run_notify_bg(notify_po_confirmed, plan_id, line_id=line_id)
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    if origin == "queue":
-        return await buy_plans_orders_partial(request, user, db)
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/lines/{line_id}/resource", response_class=HTMLResponse)
-async def buy_plan_resource_line_partial(
-    request: Request,
-    plan_id: int,
-    line_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Re-source a line whose vendor PO was cancelled.
-
-    Records the cancellation (vendor performance), marks the offer sold + the vendor
-    unavailable, drops the line into the open claim pool, and fires the URGENT backfill
-    alert to all other buyers. ``scope=plan`` re-sources the plan's other cut lines too.
-    """
-    from ..services.buyplan_notifications import notify_resource_requested, run_notify_bg
-    from ..services.buyplan_workflow import resource_line
-
-    # Per-record ownership (non-owner SALES/TRADER → 404) + PO-cutter role gate (403).
-    get_buyplan_for_user(db, user, plan_id)
-    _require_po_cutter(user)
-
-    form = await request.form()
-    reason_code = form.get("reason_code", "").strip()
-    reason_note = (form.get("reason_note") or "").strip() or None
-    scope = form.get("scope", "line")
-    origin = form.get("origin", "")
-    also_line_ids = [int(i) for i in form.getlist("also_line_ids")] if scope == "plan" else []
-
-    if not reason_code:
-        raise HTTPException(400, "A re-source reason is required")
-
-    try:
-        payload = resource_line(plan_id, line_id, reason_code, reason_note, user, db, also_line_ids=also_line_ids)
-        db.commit()
-    except ValueError as e:
-        # Log before re-raising so a real failure (e.g. an un-keyable requirement deep in
-        # the service) leaves a server trace instead of a silent, mislabeled 400.
-        logger.warning("Re-source failed for plan {} line {}: {}", plan_id, line_id, e)
-        raise HTTPException(400, str(e))
-
-    # Broadcast one urgent alert PER re-sourced line (scope=plan re-sources siblings too,
-    # and each pooled line needs its own claim).
-    for resourced in payload["resourced_lines"]:
-        await run_notify_bg(
-            notify_resource_requested, plan_id, line_id=resourced["line_id"], actor_id=user.id, reason=reason_code
-        )
-
-    if origin == "resource":
-        return await buy_plans_resource_partial(request, user, db)
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/lines/{line_id}/claim", response_class=HTMLResponse)
-async def buy_plan_claim_line_partial(
-    request: Request,
-    plan_id: int,
-    line_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Claim an open-pool (RESOURCING) line. First-to-claim wins.
-
-    No per-record ownership gate: the open pool is intentionally claimable by ANY active
-    PO-cutter regardless of who owns the parent requisition. The lost race → 409.
-    """
-    from ..services.buyplan_workflow import claim_line
-
-    _require_po_cutter(user)
-
-    form = await request.form()
-    origin = form.get("origin", "")
-
-    try:
-        claim_line(plan_id, line_id, user, db)
-        db.commit()
-    except ValueError as e:
-        logger.info("Claim lost/invalid for plan {} line {} by {}: {}", plan_id, line_id, user.id, e)
-        raise HTTPException(409, str(e))
-
-    if origin == "resource":
-        return await buy_plans_resource_partial(request, user, db)
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/lines/{line_id}/verify-po", response_class=HTMLResponse)
-async def buy_plan_verify_po_partial(
-    request: Request,
-    plan_id: int,
-    line_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Ops verifies PO — returns refreshed detail."""
-    from ..services.buyplan_notifications import (
-        notify_completed,
-        notify_po_rejected,
-        run_notify_bg,
-    )
-    from ..services.buyplan_workflow import check_completion, verify_po
-
-    form = await request.form()
-    action = form.get("action", "approve")
-    origin = form.get("origin", "")
-
-    try:
-        verify_po(plan_id, line_id, action, user, db, rejection_note=form.get("rejection_note"))
-        db.commit()
-        if action == "reject":
-            await run_notify_bg(notify_po_rejected, plan_id, line_id=line_id)
-        updated = check_completion(plan_id, db)
-        if updated and updated.status == BuyPlanStatus.COMPLETED:
-            db.commit()
-            await run_notify_bg(notify_completed, plan_id)
-    except (ValueError, PermissionError) as e:
-        raise HTTPException(400, str(e))
-
-    if origin == "supervise":
-        return _render_supervise_body(request, user, db)
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/lines/{line_id}/issue", response_class=HTMLResponse)
-async def buy_plan_flag_issue_partial(
-    request: Request,
-    plan_id: int,
-    line_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Buyer flags issue on a line — returns refreshed detail."""
-    from ..services.buyplan_workflow import flag_line_issue
-
-    # Per-record ownership: non-owner SALES/TRADER → 404 before any mutation.
-    get_buyplan_for_user(db, user, plan_id)
-
-    form = await request.form()
-    issue_type = form.get("issue_type", "other")
-    note = form.get("note", "")
-
-    try:
-        flag_line_issue(plan_id, line_id, issue_type, user, db, note=note)
-        db.commit()
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/cancel", response_class=HTMLResponse)
-async def buy_plan_cancel_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Cancel a buy plan — delegates to the service (line cascade + notification)."""
-    from ..services.buyplan_notifications import notify_cancelled, run_notify_bg
-    from ..services.buyplan_workflow import cancel_buy_plan
-
-    # Per-record ownership: non-owner SALES/TRADER → 404 before any mutation.
-    get_buyplan_for_user(db, user, plan_id)
-
-    form = await request.form()
-    try:
-        plan = cancel_buy_plan(plan_id, user, db, reason=form.get("reason"))
-        db.commit()
-        await run_notify_bg(notify_cancelled, plan.id)
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
 # ── Settings: Ops verification group ─────────────────────────────────
 
 
@@ -4832,791 +2066,6 @@ async def settings_scorecard_tab(
     if request.headers.get("HX-Request") == "true" and request.headers.get("HX-Trigger-Name") == "time_range":
         return template_response("htmx/partials/settings/_scorecard_table.html", ctx)
     return template_response("htmx/partials/settings/scorecard.html", ctx)
-
-
-@router.post("/v2/partials/buy-plans/{plan_id}/reset", response_class=HTMLResponse)
-async def buy_plan_reset_partial(
-    request: Request,
-    plan_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Reset halted/cancelled plan to draft — returns refreshed detail."""
-    from ..services.buyplan_workflow import reset_buy_plan_to_draft
-
-    # Per-record ownership: non-owner SALES/TRADER → 404 before any mutation.
-    get_buyplan_for_user(db, user, plan_id)
-
-    try:
-        reset_buy_plan_to_draft(plan_id, user, db)
-        db.commit()
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    return await buy_plan_detail_partial(request, plan_id, user, db)
-
-
-# ── Sourcing partials ──────────────────────────────────────────────────
-
-
-@router.get("/v2/sourcing/{requirement_id}", response_class=HTMLResponse)
-async def v2_sourcing_page(request: Request, requirement_id: int, db: Session = Depends(get_db)):
-    """Full page load for sourcing results."""
-    user = get_user(request, db)
-    if not user:
-        return template_response(
-            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
-        )
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["partial_url"] = f"/v2/partials/sourcing/{requirement_id}"
-    return page_response(ctx)
-
-
-@router.get("/v2/sourcing/leads/{lead_id}", response_class=HTMLResponse)
-async def v2_lead_detail_page(request: Request, lead_id: int, db: Session = Depends(get_db)):
-    """Full page load for lead detail."""
-    user = get_user(request, db)
-    if not user:
-        return template_response(
-            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
-        )
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["partial_url"] = f"/v2/partials/sourcing/leads/{lead_id}"
-    return page_response(ctx)
-
-
-@router.get("/v2/partials/sourcing/{requirement_id}/stream")
-async def sourcing_stream(
-    request: Request,
-    requirement_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """SSE endpoint for sourcing search progress.
-
-    Streams per-source completion events as connectors finish searching.
-    Client connects via hx-ext="sse" sse-connect attribute.
-    Channel: sourcing:{requirement_id}
-    """
-    from sse_starlette.sse import EventSourceResponse
-
-    from ..services.sse_broker import broker
-
-    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
-    if not req:
-        raise HTTPException(404, "Requirement not found")
-
-    async def event_generator():
-        async for msg in broker.listen(f"sourcing:{requirement_id}"):
-            if await request.is_disconnected():
-                break
-            yield {
-                "event": msg["event"],
-                "data": msg["data"],
-            }
-
-    return EventSourceResponse(event_generator())
-
-
-@router.post("/v2/partials/sourcing/{requirement_id}/search", response_class=HTMLResponse)
-async def sourcing_search_trigger(
-    request: Request,
-    requirement_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Trigger multi-source search for a requirement.
-
-    Runs connectors in parallel, publishes SSE events per source completion, syncs leads
-    on completion, returns redirect to sourcing results.
-    """
-    import asyncio
-    import json
-
-    from ..services.sse_broker import broker
-
-    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
-    if not req:
-        raise HTTPException(404, "Requirement not found")
-    # Search triggers connector SPEND + cross-owner disclosure — scope to the owner.
-    require_requisition_access(db, req.requisition_id, user, label="Requirement")
-
-    mpn = req.primary_mpn or ""
-    sources = ["brokerbin", "nexar", "digikey", "mouser", "oemsecrets", "element14"]
-    channel = f"sourcing:{requirement_id}"
-    all_sightings = []
-
-    async def search_source(source_name):
-        start_t = time.time()
-        try:
-            from ..search_service import quick_search_mpn
-
-            raw = await quick_search_mpn(mpn, db)
-            results = raw if isinstance(raw, list) else raw.get("sightings", [])
-            elapsed = int((time.time() - start_t) * 1000)
-            count = len(results) if results else 0
-            await broker.publish(
-                channel,
-                "source-complete",
-                json.dumps({"source": source_name, "count": count, "elapsed_ms": elapsed, "status": "done"}),
-            )
-            return results or []
-        except Exception as exc:
-            elapsed = int((time.time() - start_t) * 1000)
-            logger.error("Sourcing search failed for {} on {}: {}", mpn, source_name, exc)
-            await broker.publish(
-                channel,
-                "source-complete",
-                json.dumps(
-                    {"source": source_name, "count": 0, "elapsed_ms": elapsed, "status": "failed", "error": str(exc)}
-                ),
-            )
-            return []
-
-    results_by_source = await asyncio.gather(*[search_source(s) for s in sources], return_exceptions=True)
-
-    for source_results in results_by_source:
-        if isinstance(source_results, list):
-            all_sightings.extend(source_results)
-
-    await broker.publish(
-        channel, "search-complete", json.dumps({"total": len(all_sightings), "requirement_id": requirement_id})
-    )
-
-    return HTMLResponse(status_code=200, headers={"HX-Redirect": f"/v2/sourcing/{requirement_id}"})
-
-
-@router.get("/v2/partials/sourcing/{requirement_id}", response_class=HTMLResponse)
-async def sourcing_results_partial(
-    request: Request,
-    requirement_id: int,
-    confidence: str = "",
-    safety: str = "",
-    freshness: str = "",
-    source: str = "",
-    status: str = "",
-    contactability: str = "",
-    corroborated: str = "",
-    sort: str = "best",
-    page: int = Query(1, ge=1),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return sourcing results with lead cards for a requirement.
-
-    Supports filtering by confidence band, safety band, freshness window, source type,
-    buyer status, contactability, and corroboration. Sorts by best overall (default),
-    freshest, safest, easiest to contact, or most proven.
-    """
-    from ..models.sourcing_lead import SourcingLead
-
-    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
-    if not req:
-        raise HTTPException(404, "Requirement not found")
-
-    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
-
-    if confidence:
-        bands = [b.strip() for b in confidence.split(",")]
-        query = query.filter(SourcingLead.confidence_band.in_(bands))
-    if safety:
-        bands = [b.strip() for b in safety.split(",")]
-        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
-    if freshness and freshness != "all":
-        now = datetime.now(timezone.utc)
-        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
-        if freshness in cutoffs:
-            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
-    if source:
-        sources_list = [s.strip() for s in source.split(",")]
-        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
-    if status and status != "all":
-        statuses = [s.strip() for s in status.split(",")]
-        query = query.filter(SourcingLead.buyer_status.in_(statuses))
-    if contactability == "has_email":
-        query = query.filter(SourcingLead.contact_email.isnot(None))
-    elif contactability == "has_phone":
-        query = query.filter(SourcingLead.contact_phone.isnot(None))
-    if corroborated == "yes":
-        query = query.filter(SourcingLead.corroborated.is_(True))
-    elif corroborated == "no":
-        query = query.filter(SourcingLead.corroborated.is_(False))
-
-    sort_map = {
-        "best": [SourcingLead.confidence_score.desc()],
-        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
-        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
-        "contact": [SourcingLead.contactability_score.desc().nullslast()],
-        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
-    }
-    for col in sort_map.get(sort, sort_map["best"]):
-        query = query.order_by(col)
-
-    total = query.count()
-    per_page = 24
-    leads = query.offset((page - 1) * per_page).limit(per_page).all()
-
-    lead_sighting_data = {}
-    if leads:
-        for lead in leads:
-            best_sighting = (
-                db.query(Sighting)
-                .filter(
-                    Sighting.requirement_id == requirement_id,
-                    Sighting.vendor_name_normalized == lead.vendor_name_normalized,
-                )
-                .order_by(Sighting.created_at.desc().nullslast())
-                .first()
-            )
-            if best_sighting:
-                lead_sighting_data[lead.id] = {
-                    "qty_available": best_sighting.qty_available,
-                    "unit_price": best_sighting.unit_price,
-                }
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update(
-        {
-            "requirement": req,
-            "leads": leads,
-            "lead_sighting_data": lead_sighting_data,
-            "total": total,
-            "page": page,
-            "total_pages": max(1, (total + per_page - 1) // per_page),
-            "per_page": per_page,
-            "f_confidence": confidence,
-            "f_safety": safety,
-            "f_freshness": freshness,
-            "f_source": source,
-            "f_status": status,
-            "f_contactability": contactability,
-            "f_corroborated": corroborated,
-            "f_sort": sort,
-        }
-    )
-    return template_response("htmx/partials/sourcing/results.html", ctx)
-
-
-@router.get("/v2/partials/sourcing/leads/{lead_id}", response_class=HTMLResponse)
-async def lead_detail_partial(
-    request: Request,
-    lead_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return lead detail as HTML partial.
-
-    Loads the SourcingLead, its evidence (sorted by confidence_impact desc), groups
-    evidence by source category, fetches vendor card and best sighting.
-    """
-    from ..models.sourcing_lead import LeadEvidence, SourcingLead
-    from ..services.sourcing_leads import _source_category
-
-    lead = db.query(SourcingLead).filter(SourcingLead.id == lead_id).first()
-    if not lead:
-        raise HTTPException(404, "Lead not found")
-
-    evidence = (
-        db.query(LeadEvidence)
-        .filter(LeadEvidence.lead_id == lead.id)
-        .order_by(LeadEvidence.confidence_impact.desc().nullslast())
-        .all()
-    )
-
-    evidence_by_category = {}
-    for ev in evidence:
-        cat = _source_category(ev.source_type)
-        evidence_by_category.setdefault(cat, []).append(ev)
-
-    category_labels = {
-        "api": "API",
-        "marketplace": "Marketplace",
-        "salesforce_history": "Salesforce History",
-        "avail_history": "Avail History",
-        "web_ai": "Web / AI",
-        "safety_review": "Safety Review",
-        "buyer_feedback": "Buyer Feedback",
-    }
-
-    requirement = db.query(Requirement).filter(Requirement.id == lead.requirement_id).first()
-
-    vendor_card = lead.vendor_card
-
-    best_sighting = (
-        db.query(Sighting)
-        .filter(
-            Sighting.requirement_id == lead.requirement_id,
-            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
-        )
-        .order_by(Sighting.created_at.desc().nullslast())
-        .first()
-    )
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update(
-        {
-            "lead": lead,
-            "evidence": evidence,
-            "evidence_by_category": evidence_by_category,
-            "category_labels": category_labels,
-            "requirement": requirement,
-            "vendor_card": vendor_card,
-            "best_sighting": best_sighting,
-        }
-    )
-    return template_response("htmx/partials/sourcing/lead_detail.html", ctx)
-
-
-@router.post("/v2/partials/sourcing/leads/{lead_id}/status", response_class=HTMLResponse)
-async def lead_status_update(
-    request: Request,
-    lead_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Update lead buyer status.
-
-    Returns updated lead card when called from results view (for OOB swap), or updated
-    lead detail when called from lead detail page.
-    """
-    from ..models.sourcing_lead import SourcingLead
-    from ..services.sourcing_leads import update_lead_status
-
-    _lead = db.get(SourcingLead, lead_id)
-    if not _lead:
-        raise HTTPException(404, "Lead not found")
-    require_requisition_access(db, _lead.requisition_id, user, label="Lead")
-
-    form = await request.form()
-    status_val = form.get("status", "").strip()
-    note = form.get("note", "").strip() or None
-
-    try:
-        lead = update_lead_status(
-            db,
-            lead_id,
-            status_val,
-            note=note,
-            actor_user_id=user.id,
-        )
-    except ValueError as e:
-        raise HTTPException(400, str(e))
-
-    if not lead:
-        raise HTTPException(404, "Lead not found")
-
-    hx_target = request.headers.get("HX-Target", "")
-    referer = request.headers.get("HX-Current-URL", "")
-
-    # Workspace panel context: return updated panel detail
-    if hx_target == "split-right-sourcing":
-        return await lead_panel_partial(request, lead_id, user, db)
-
-    # Full-page lead detail context
-    if "/leads/" in referer:
-        return await lead_detail_partial(request, lead_id, user, db)
-
-    # Workspace lead row context: return updated lead row
-    if hx_target.startswith("lead-row-"):
-        best_sighting = (
-            db.query(Sighting)
-            .filter(
-                Sighting.requirement_id == lead.requirement_id,
-                Sighting.vendor_name_normalized == lead.vendor_name_normalized,
-            )
-            .order_by(Sighting.created_at.desc().nullslast())
-            .first()
-        )
-        lead_sighting_data = {}
-        if best_sighting:
-            lead_sighting_data[lead.id] = {
-                "qty_available": best_sighting.qty_available,
-                "unit_price": best_sighting.unit_price,
-            }
-        ctx = _base_ctx(request, user, "requisitions")
-        ctx.update({"lead": lead, "lead_sighting_data": lead_sighting_data, "selected_lead_id": 0})
-        return template_response("htmx/partials/sourcing/lead_row.html", ctx)
-
-    # Default: card view (results grid)
-    best_sighting = (
-        db.query(Sighting)
-        .filter(
-            Sighting.requirement_id == lead.requirement_id,
-            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
-        )
-        .order_by(Sighting.created_at.desc().nullslast())
-        .first()
-    )
-    lead_sighting_data = {}
-    if best_sighting:
-        lead_sighting_data[lead.id] = {
-            "qty_available": best_sighting.qty_available,
-            "unit_price": best_sighting.unit_price,
-        }
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update({"lead": lead, "lead_sighting_data": lead_sighting_data})
-    return template_response("htmx/partials/sourcing/lead_card.html", ctx)
-
-
-@router.post("/v2/partials/sourcing/leads/{lead_id}/feedback", response_class=HTMLResponse)
-async def lead_feedback(
-    request: Request,
-    lead_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Add feedback event to a lead without changing status.
-
-    Returns updated lead detail.
-    """
-    from ..models.sourcing_lead import SourcingLead
-    from ..services.sourcing_leads import append_lead_feedback
-
-    _lead = db.get(SourcingLead, lead_id)
-    if not _lead:
-        raise HTTPException(404, "Lead not found")
-    require_requisition_access(db, _lead.requisition_id, user, label="Lead")
-
-    form = await request.form()
-    note = form.get("note", "").strip() or None
-    reason_code = form.get("reason_code", "").strip() or None
-    contact_method = form.get("contact_method", "").strip() or None
-
-    lead = append_lead_feedback(
-        db,
-        lead_id,
-        note=note,
-        reason_code=reason_code,
-        contact_method=contact_method,
-        actor_user_id=user.id,
-    )
-    if not lead:
-        raise HTTPException(404, "Lead not found")
-
-    return await lead_detail_partial(request, lead_id, user, db)
-
-
-# ── Sourcing workspace (split-panel) ─────────────────────────────────
-
-
-@router.get("/v2/sourcing/{requirement_id}/workspace", response_class=HTMLResponse)
-async def v2_sourcing_workspace_page(request: Request, requirement_id: int, db: Session = Depends(get_db)):
-    """Full page load for sourcing workspace (split-panel view)."""
-    user = get_user(request, db)
-    if not user:
-        return template_response(
-            "htmx/login.html", {"request": request, "password_login": _password_login_enabled(), **_vite_assets()}
-        )
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx["partial_url"] = f"/v2/partials/sourcing/{requirement_id}/workspace"
-    return page_response(ctx)
-
-
-@router.get("/v2/partials/sourcing/{requirement_id}/workspace", response_class=HTMLResponse)
-async def sourcing_workspace_partial(
-    request: Request,
-    requirement_id: int,
-    confidence: str = "",
-    safety: str = "",
-    freshness: str = "",
-    source: str = "",
-    status: str = "",
-    contactability: str = "",
-    corroborated: str = "",
-    sort: str = "best",
-    page: int = Query(1, ge=1),
-    lead: int = Query(0, ge=0),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return sourcing workspace split-panel layout.
-
-    Reuses the same filtering/sorting logic as sourcing_results_partial but renders lead
-    rows in a split-panel instead of card grid. Optional lead=ID param pre-selects a
-    lead in the right panel.
-    """
-    from ..models.sourcing_lead import SourcingLead
-
-    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
-    if not req:
-        raise HTTPException(404, "Requirement not found")
-
-    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
-
-    if confidence:
-        bands = [b.strip() for b in confidence.split(",")]
-        query = query.filter(SourcingLead.confidence_band.in_(bands))
-    if safety:
-        bands = [b.strip() for b in safety.split(",")]
-        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
-    if freshness and freshness != "all":
-        now = datetime.now(timezone.utc)
-        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
-        if freshness in cutoffs:
-            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
-    if source:
-        sources_list = [s.strip() for s in source.split(",")]
-        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
-    if status and status != "all":
-        statuses = [s.strip() for s in status.split(",")]
-        query = query.filter(SourcingLead.buyer_status.in_(statuses))
-    if contactability == "has_email":
-        query = query.filter(SourcingLead.contact_email.isnot(None))
-    elif contactability == "has_phone":
-        query = query.filter(SourcingLead.contact_phone.isnot(None))
-    if corroborated == "yes":
-        query = query.filter(SourcingLead.corroborated.is_(True))
-    elif corroborated == "no":
-        query = query.filter(SourcingLead.corroborated.is_(False))
-
-    sort_map = {
-        "best": [SourcingLead.confidence_score.desc()],
-        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
-        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
-        "contact": [SourcingLead.contactability_score.desc().nullslast()],
-        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
-    }
-    for col in sort_map.get(sort, sort_map["best"]):
-        query = query.order_by(col)
-
-    total = query.count()
-    per_page = 24
-    leads = query.offset((page - 1) * per_page).limit(per_page).all()
-
-    lead_sighting_data = {}
-    if leads:
-        for ld in leads:
-            best_sighting = (
-                db.query(Sighting)
-                .filter(
-                    Sighting.requirement_id == requirement_id,
-                    Sighting.vendor_name_normalized == ld.vendor_name_normalized,
-                )
-                .order_by(Sighting.created_at.desc().nullslast())
-                .first()
-            )
-            if best_sighting:
-                lead_sighting_data[ld.id] = {
-                    "qty_available": best_sighting.qty_available,
-                    "unit_price": best_sighting.unit_price,
-                }
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update(
-        {
-            "requirement": req,
-            "leads": leads,
-            "lead_sighting_data": lead_sighting_data,
-            "total": total,
-            "page": page,
-            "total_pages": max(1, (total + per_page - 1) // per_page),
-            "per_page": per_page,
-            "f_confidence": confidence,
-            "f_safety": safety,
-            "f_freshness": freshness,
-            "f_source": source,
-            "f_status": status,
-            "f_contactability": contactability,
-            "f_corroborated": corroborated,
-            "f_sort": sort,
-            "selected_lead_id": lead if lead else 0,
-        }
-    )
-    return template_response("htmx/partials/sourcing/workspace.html", ctx)
-
-
-@router.get("/v2/partials/sourcing/{requirement_id}/workspace-list", response_class=HTMLResponse)
-async def sourcing_workspace_list_partial(
-    request: Request,
-    requirement_id: int,
-    confidence: str = "",
-    safety: str = "",
-    freshness: str = "",
-    source: str = "",
-    status: str = "",
-    contactability: str = "",
-    corroborated: str = "",
-    sort: str = "best",
-    page: int = Query(1, ge=1),
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return just the lead list rows for the workspace left panel.
-
-    Used when filters change — swaps only #lead-list-content without touching the right
-    panel or overall layout.
-    """
-    from ..models.sourcing_lead import SourcingLead
-
-    req = db.query(Requirement).filter(Requirement.id == requirement_id).first()
-    if not req:
-        raise HTTPException(404, "Requirement not found")
-
-    query = db.query(SourcingLead).filter(SourcingLead.requirement_id == requirement_id)
-
-    if confidence:
-        bands = [b.strip() for b in confidence.split(",")]
-        query = query.filter(SourcingLead.confidence_band.in_(bands))
-    if safety:
-        bands = [b.strip() for b in safety.split(",")]
-        query = query.filter(SourcingLead.vendor_safety_band.in_(bands))
-    if freshness and freshness != "all":
-        now = datetime.now(timezone.utc)
-        cutoffs = {"24h": timedelta(hours=24), "7d": timedelta(days=7), "30d": timedelta(days=30)}
-        if freshness in cutoffs:
-            query = query.filter(SourcingLead.source_last_seen_at >= now - cutoffs[freshness])
-    if source:
-        sources_list = [s.strip() for s in source.split(",")]
-        query = query.filter(SourcingLead.primary_source_type.in_(sources_list))
-    if status and status != "all":
-        statuses = [s.strip() for s in status.split(",")]
-        query = query.filter(SourcingLead.buyer_status.in_(statuses))
-    if contactability == "has_email":
-        query = query.filter(SourcingLead.contact_email.isnot(None))
-    elif contactability == "has_phone":
-        query = query.filter(SourcingLead.contact_phone.isnot(None))
-    if corroborated == "yes":
-        query = query.filter(SourcingLead.corroborated.is_(True))
-    elif corroborated == "no":
-        query = query.filter(SourcingLead.corroborated.is_(False))
-
-    sort_map = {
-        "best": [SourcingLead.confidence_score.desc()],
-        "freshest": [SourcingLead.source_last_seen_at.desc().nullslast()],
-        "safest": [SourcingLead.vendor_safety_score.desc().nullslast()],
-        "contact": [SourcingLead.contactability_score.desc().nullslast()],
-        "proven": [SourcingLead.historical_success_score.desc().nullslast()],
-    }
-    for col in sort_map.get(sort, sort_map["best"]):
-        query = query.order_by(col)
-
-    total = query.count()
-    per_page = 24
-    leads = query.offset((page - 1) * per_page).limit(per_page).all()
-
-    lead_sighting_data = {}
-    if leads:
-        for ld in leads:
-            best_sighting = (
-                db.query(Sighting)
-                .filter(
-                    Sighting.requirement_id == requirement_id,
-                    Sighting.vendor_name_normalized == ld.vendor_name_normalized,
-                )
-                .order_by(Sighting.created_at.desc().nullslast())
-                .first()
-            )
-            if best_sighting:
-                lead_sighting_data[ld.id] = {
-                    "qty_available": best_sighting.qty_available,
-                    "unit_price": best_sighting.unit_price,
-                }
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update(
-        {
-            "requirement": req,
-            "leads": leads,
-            "lead_sighting_data": lead_sighting_data,
-            "total": total,
-            "page": page,
-            "total_pages": max(1, (total + per_page - 1) // per_page),
-            "per_page": per_page,
-            "f_confidence": confidence,
-            "f_safety": safety,
-            "f_freshness": freshness,
-            "f_source": source,
-            "f_status": status,
-            "f_contactability": contactability,
-            "f_corroborated": corroborated,
-            "f_sort": sort,
-            "selected_lead_id": 0,
-        }
-    )
-
-    html_parts = []
-    if leads:
-        for ld in leads:
-            rendered = templates.get_template("htmx/partials/sourcing/lead_row.html").render({**ctx, "lead": ld})
-            html_parts.append(rendered)
-    else:
-        html_parts.append(
-            '<div class="flex flex-col items-center justify-center py-12 text-gray-400">'
-            '<p class="text-sm">No leads found</p></div>'
-        )
-
-    return HTMLResponse("".join(html_parts))
-
-
-@router.get("/v2/partials/sourcing/leads/{lead_id}/panel", response_class=HTMLResponse)
-async def lead_panel_partial(
-    request: Request,
-    lead_id: int,
-    user: User = Depends(require_user),
-    db: Session = Depends(get_db),
-):
-    """Return condensed lead detail for workspace right panel.
-
-    Same data as lead_detail_partial but uses lead_panel.html template (no breadcrumb,
-    collapsible sections, denser layout). Includes OOB swap of the lead row in the left
-    panel for highlight update.
-    """
-    from ..models.sourcing_lead import LeadEvidence, SourcingLead
-    from ..services.sourcing_leads import _source_category
-
-    lead = db.query(SourcingLead).filter(SourcingLead.id == lead_id).first()
-    if not lead:
-        raise HTTPException(404, "Lead not found")
-
-    evidence = (
-        db.query(LeadEvidence)
-        .filter(LeadEvidence.lead_id == lead.id)
-        .order_by(LeadEvidence.confidence_impact.desc().nullslast())
-        .all()
-    )
-
-    evidence_by_category = {}
-    for ev in evidence:
-        cat = _source_category(ev.source_type)
-        evidence_by_category.setdefault(cat, []).append(ev)
-
-    category_labels = {
-        "api": "API",
-        "marketplace": "Marketplace",
-        "salesforce_history": "Salesforce History",
-        "avail_history": "Avail History",
-        "web_ai": "Web / AI",
-        "safety_review": "Safety Review",
-        "buyer_feedback": "Buyer Feedback",
-    }
-
-    requirement = db.query(Requirement).filter(Requirement.id == lead.requirement_id).first()
-    vendor_card = lead.vendor_card
-
-    best_sighting = (
-        db.query(Sighting)
-        .filter(
-            Sighting.requirement_id == lead.requirement_id,
-            Sighting.vendor_name_normalized == lead.vendor_name_normalized,
-        )
-        .order_by(Sighting.created_at.desc().nullslast())
-        .first()
-    )
-
-    ctx = _base_ctx(request, user, "requisitions")
-    ctx.update(
-        {
-            "lead": lead,
-            "evidence": evidence,
-            "evidence_by_category": evidence_by_category,
-            "category_labels": category_labels,
-            "requirement": requirement,
-            "vendor_card": vendor_card,
-            "best_sighting": best_sighting,
-        }
-    )
-    return template_response("htmx/partials/sourcing/lead_panel.html", ctx)
 
 
 # ── Materials partials ────────────────────────────────────────────────
